% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.fastNaiveBayes.R
\name{print.fastNaiveBayes}
\alias{print.fastNaiveBayes}
\title{Prints fastNaiveBayes object}
\usage{
\method{print}{fastNaiveBayes}(x, ...)
}
\arguments{
\item{x}{a fastNaiveBayes object}

\item{...}{not used.}
}
\description{
Prints fastNaiveBayes object
}
