% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tweets}
\alias{tweets}
\title{This data originally came from Crowdflower's Data for Everyone library.}
\format{A data frame with 14640 rows and 2 columns
\describe{
   \item{airline_sentiment}{sentiment, as either 'negative','neutral','positive'}
   \item{text}{raw text, character}
}}
\source{
\url{https://www.figure-eight.com/data-for-everyone/}
}
\usage{
tweets
}
\description{
As the original source says,
}
\details{
"A sentiment analysis job about the problems of each major U.S. airline.
Twitter data was scraped from February of 2015 and contributors were
asked to first classify positive, negative, and neutral tweets, followed
by categorizing negative reasons (such as "late flight" or "rude service")."

The data provided here is an altered version of the original source.
}
\keyword{datasets}
