% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tweetsDTM}
\alias{tweetsDTM}
\title{This data originally came from Crowdflower's Data for Everyone library.}
\format{A data frame with 14640 rows and 2186 columns
\describe{
   \item{airline_sentiment}{sentiment, as either 'negative','neutral','positive'}
   \item{said}{Frequency count, numeric}
   \item{virginamerica}{Frequency count, numeric}
   \item{commerci}{Frequency count, numeric}
   \item{experi}{Frequency count, numeric}
   \item{plus}{Frequency count, numeric}
   \item{anoth}{Frequency count, numeric}
   \item{mean}{Frequency count, numeric}
   \item{must}{Frequency count, numeric}
   \item{need}{Frequency count, numeric}
   \item{take}{Frequency count, numeric}
   \item{today}{Frequency count, numeric}
   \item{trip}{Frequency count, numeric}
   \item{amp}{Frequency count, numeric}
   \item{blast}{Frequency count, numeric}
   \item{entertain}{Frequency count, numeric}
   \item{face}{Frequency count, numeric}
   \item{guest}{Frequency count, numeric}
   \item{littl}{Frequency count, numeric}
   \item{realli}{Frequency count, numeric}
   \item{recours}{Frequency count, numeric}
   \item{bad}{Frequency count, numeric}
   \item{big}{Frequency count, numeric}
   \item{thing}{Frequency count, numeric}
   \item{fli}{Frequency count, numeric}
   \item{flight}{Frequency count, numeric}
   \item{pay}{Frequency count, numeric}
   \item{play}{Frequency count, numeric}
   \item{seat}{Frequency count, numeric}
   \item{serious}{Frequency count, numeric}
   \item{away}{Frequency count, numeric}
   \item{everi}{Frequency count, numeric}
   \item{near}{Frequency count, numeric}
   \item{time}{Frequency count, numeric}
   \item{wonut}{Frequency count, numeric}
   \item{yes}{Frequency count, numeric}
   \item{hat}{Frequency count, numeric}
   \item{miss}{Frequency count, numeric}
   \item{opportun}{Frequency count, numeric}
   \item{without}{Frequency count, numeric}
   \item{now}{Frequency count, numeric}
   \item{well}{Frequency count, numeric}
   \item{amaz}{Frequency count, numeric}
   \item{arriv}{Frequency count, numeric}
   \item{earli}{Frequency count, numeric}
   \item{good}{Frequency count, numeric}
   \item{hour}{Frequency count, numeric}
   \item{among}{Frequency count, numeric}
   \item{caus}{Frequency count, numeric}
   \item{death}{Frequency count, numeric}
   \item{know}{Frequency count, numeric}
   \item{lead}{Frequency count, numeric}
   \item{second}{Frequency count, numeric}
   \item{suicid}{Frequency count, numeric}
   \item{teen}{Frequency count, numeric}
   \item{better}{Frequency count, numeric}
   \item{much}{Frequency count, numeric}
   \item{pretti}{Frequency count, numeric}
   \item{alreadi}{Frequency count, numeric}
   \item{australia}{Frequency count, numeric}
   \item{deal}{Frequency count, numeric}
   \item{even}{Frequency count, numeric}
   \item{gone}{Frequency count, numeric}
   \item{great}{Frequency count, numeric}
   \item{think}{Frequency count, numeric}
   \item{yet}{Frequency count, numeric}
   \item{fabul}{Frequency count, numeric}
   \item{sky}{Frequency count, numeric}
   \item{stress}{Frequency count, numeric}
   \item{travel}{Frequency count, numeric}
   \item{thank}{Frequency count, numeric}
   \item{mia}{Frequency count, numeric}
   \item{schedul}{Frequency count, numeric}
   \item{still}{Frequency count, numeric}
   \item{america}{Frequency count, numeric}
   \item{countri}{Frequency count, numeric}
   \item{cross}{Frequency count, numeric}
   \item{excit}{Frequency count, numeric}
   \item{first}{Frequency count, numeric}
   \item{heard}{Frequency count, numeric}
   \item{lax}{Frequency count, numeric}
   \item{mco}{Frequency count, numeric}
   \item{noth}{Frequency count, numeric}
   \item{virgin}{Frequency count, numeric}
   \item{due}{Frequency count, numeric}
   \item{either}{Frequency count, numeric}
   \item{flew}{Frequency count, numeric}
   \item{fulli}{Frequency count, numeric}
   \item{help}{Frequency count, numeric}
   \item{larg}{Frequency count, numeric}
   \item{last}{Frequency count, numeric}
   \item{nyc}{Frequency count, numeric}
   \item{sfo}{Frequency count, numeric}
   \item{side}{Frequency count, numeric}
   \item{sit}{Frequency count, numeric}
   \item{two}{Frequency count, numeric}
   \item{week}{Frequency count, numeric}
   \item{cent}{Frequency count, numeric}
   \item{uuo}{Frequency count, numeric}
   \item{awesom}{Frequency count, numeric}
   \item{pleas}{Frequency count, numeric}
   \item{want}{Frequency count, numeric}
   \item{avail}{Frequency count, numeric}
   \item{carrier}{Frequency count, numeric}
   \item{fare}{Frequency count, numeric}
   \item{may}{Frequency count, numeric}
   \item{select}{Frequency count, numeric}
   \item{three}{Frequency count, numeric}
   \item{love}{Frequency count, numeric}
   \item{brand}{Frequency count, numeric}
   \item{feel}{Frequency count, numeric}
   \item{anytim}{Frequency count, numeric}
   \item{make}{Frequency count, numeric}
   \item{non}{Frequency count, numeric}
   \item{soon}{Frequency count, numeric}
   \item{stop}{Frequency count, numeric}
   \item{will}{Frequency count, numeric}
   \item{free}{Frequency count, numeric}
   \item{friend}{Frequency count, numeric}
   \item{gave}{Frequency count, numeric}
   \item{guy}{Frequency count, numeric}
   \item{internet}{Frequency count, numeric}
   \item{mess}{Frequency count, numeric}
   \item{reserv}{Frequency count, numeric}
   \item{appli}{Frequency count, numeric}
   \item{call}{Frequency count, numeric}
   \item{email}{Frequency count, numeric}
   \item{match}{Frequency count, numeric}
   \item{program}{Frequency count, numeric}
   \item{respons}{Frequency count, numeric}
   \item{status}{Frequency count, numeric}
   \item{abl}{Frequency count, numeric}
   \item{anyth}{Frequency count, numeric}
   \item{eat}{Frequency count, numeric}
   \item{fail}{Frequency count, numeric}
   \item{food}{Frequency count, numeric}
   \item{happen}{Frequency count, numeric}
   \item{hrs}{Frequency count, numeric}
   \item{least}{Frequency count, numeric}
   \item{next}{Frequency count, numeric}
   \item{option}{Frequency count, numeric}
   \item{say}{Frequency count, numeric}
   \item{site}{Frequency count, numeric}
   \item{togeth}{Frequency count, numeric}
   \item{worri}{Frequency count, numeric}
   \item{air}{Frequency count, numeric}
   \item{cold}{Frequency count, numeric}
   \item{get}{Frequency count, numeric}
   \item{ewr}{Frequency count, numeric}
   \item{eye}{Frequency count, numeric}
   \item{middl}{Frequency count, numeric}
   \item{red}{Frequency count, numeric}
   \item{add}{Frequency count, numeric}
   \item{birthday}{Frequency count, numeric}
   \item{book}{Frequency count, numeric}
   \item{cool}{Frequency count, numeric}
   \item{elev}{Frequency count, numeric}
   \item{enter}{Frequency count, numeric}
   \item{just}{Frequency count, numeric}
   \item{name}{Frequency count, numeric}
   \item{problem}{Frequency count, numeric}
   \item{club}{Frequency count, numeric}
   \item{current}{Frequency count, numeric}
   \item{onlin}{Frequency count, numeric}
   \item{oper}{Frequency count, numeric}
   \item{post}{Frequency count, numeric}
   \item{answer}{Frequency count, numeric}
   \item{expens}{Frequency count, numeric}
   \item{headphon}{Frequency count, numeric}
   \item{iad}{Frequency count, numeric}
   \item{left}{Frequency count, numeric}
   \item{number}{Frequency count, numeric}
   \item{one}{Frequency count, numeric}
   \item{await}{Frequency count, numeric}
   \item{phone}{Frequency count, numeric}
   \item{prefer}{Frequency count, numeric}
   \item{return}{Frequency count, numeric}
   \item{use}{Frequency count, numeric}
   \item{end}{Frequency count, numeric}
   \item{hawaii}{Frequency count, numeric}
   \item{news}{Frequency count, numeric}
   \item{start}{Frequency count, numeric}
   \item{via}{Frequency count, numeric}
   \item{year}{Frequency count, numeric}
   \item{nice}{Frequency count, numeric}
   \item{takeoff}{Frequency count, numeric}
   \item{touchdown}{Frequency count, numeric}
   \item{best}{Frequency count, numeric}
   \item{ever}{Frequency count, numeric}
   \item{way}{Frequency count, numeric}
   \item{airlin}{Frequency count, numeric}
   \item{around}{Frequency count, numeric}
   \item{done}{Frequency count, numeric}
   \item{hand}{Frequency count, numeric}
   \item{can}{Frequency count, numeric}
   \item{chat}{Frequency count, numeric}
   \item{support}{Frequency count, numeric}
   \item{work}{Frequency count, numeric}
   \item{angel}{Frequency count, numeric}
   \item{beyond}{Frequency count, numeric}
   \item{sign}{Frequency count, numeric}
   \item{view}{Frequency count, numeric}
   \item{account}{Frequency count, numeric}
   \item{flyer}{Frequency count, numeric}
   \item{hard}{Frequency count, numeric}
   \item{hey}{Frequency count, numeric}
   \item{plz}{Frequency count, numeric}
   \item{upgrad}{Frequency count, numeric}
   \item{win}{Frequency count, numeric}
   \item{citi}{Frequency count, numeric}
   \item{expir}{Frequency count, numeric}
   \item{move}{Frequency count, numeric}
   \item{new}{Frequency count, numeric}
   \item{ticket}{Frequency count, numeric}
   \item{unus}{Frequency count, numeric}
   \item{dalla}{Frequency count, numeric}
   \item{feb}{Frequency count, numeric}
   \item{leav}{Frequency count, numeric}
   \item{seattl}{Frequency count, numeric}
   \item{reason}{Frequency count, numeric}
   \item{rock}{Frequency count, numeric}
   \item{dream}{Frequency count, numeric}
   \item{mind}{Frequency count, numeric}
   \item{wow}{Frequency count, numeric}
   \item{agre}{Frequency count, numeric}
   \item{carrieunderwood}{Frequency count, numeric}
   \item{ladygaga}{Frequency count, numeric}
   \item{night}{Frequency count, numeric}
   \item{oscar}{Frequency count, numeric}
   \item{ago}{Frequency count, numeric}
   \item{minut}{Frequency count, numeric}
   \item{show}{Frequency count, numeric}
   \item{suppos}{Frequency count, numeric}
   \item{websit}{Frequency count, numeric}
   \item{andrew}{Frequency count, numeric}
   \item{impress}{Frequency count, numeric}
   \item{juli}{Frequency count, numeric}
   \item{though}{Frequency count, numeric}
   \item{atlanta}{Frequency count, numeric}
   \item{wish}{Frequency count, numeric}
   \item{februari}{Frequency count, numeric}
   \item{load}{Frequency count, numeric}
   \item{never}{Frequency count, numeric}
   \item{page}{Frequency count, numeric}
   \item{sinc}{Frequency count, numeric}
   \item{thx}{Frequency count, numeric}
   \item{tri}{Frequency count, numeric}
   \item{stat}{Frequency count, numeric}
   \item{ladi}{Frequency count, numeric}
   \item{carri}{Frequency count, numeric}
   \item{let}{Frequency count, numeric}
   \item{market}{Frequency count, numeric}
   \item{song}{Frequency count, numeric}
   \item{shown}{Frequency count, numeric}
   \item{mari}{Frequency count, numeric}
   \item{sorri}{Frequency count, numeric}
   \item{beat}{Frequency count, numeric}
   \item{check}{Frequency count, numeric}
   \item{dal}{Frequency count, numeric}
   \item{dca}{Frequency count, numeric}
   \item{someon}{Frequency count, numeric}
   \item{earlier}{Frequency count, numeric}
   \item{hold}{Frequency count, numeric}
   \item{tonight}{Frequency count, numeric}
   \item{award}{Frequency count, numeric}
   \item{congrat}{Frequency count, numeric}
   \item{bag}{Frequency count, numeric}
   \item{everyth}{Frequency count, numeric}
   \item{fine}{Frequency count, numeric}
   \item{lost}{Frequency count, numeric}
   \item{card}{Frequency count, numeric}
   \item{chang}{Frequency count, numeric}
   \item{credit}{Frequency count, numeric}
   \item{fee}{Frequency count, numeric}
   \item{waiv}{Frequency count, numeric}
   \item{custom}{Frequency count, numeric}
   \item{servic}{Frequency count, numeric}
   \item{team}{Frequency count, numeric}
   \item{track}{Frequency count, numeric}
   \item{baggag}{Frequency count, numeric}
   \item{dirti}{Frequency count, numeric}
   \item{floor}{Frequency count, numeric}
   \item{game}{Frequency count, numeric}
   \item{step}{Frequency count, numeric}
   \item{tabl}{Frequency count, numeric}
   \item{crew}{Frequency count, numeric}
   \item{like}{Frequency count, numeric}
   \item{plane}{Frequency count, numeric}
   \item{ride}{Frequency count, numeric}
   \item{bill}{Frequency count, numeric}
   \item{correct}{Frequency count, numeric}
   \item{morn}{Frequency count, numeric}
   \item{sat}{Frequency count, numeric}
   \item{yall}{Frequency count, numeric}
   \item{feet}{Frequency count, numeric}
   \item{film}{Frequency count, numeric}
   \item{student}{Frequency count, numeric}
   \item{watch}{Frequency count, numeric}
   \item{differ}{Frequency count, numeric}
   \item{media}{Frequency count, numeric}
   \item{anyway}{Frequency count, numeric}
   \item{asap}{Frequency count, numeric}
   \item{human}{Frequency count, numeric}
   \item{speak}{Frequency count, numeric}
   \item{biz}{Frequency count, numeric}
   \item{jetblu}{Frequency count, numeric}
   \item{southwestair}{Frequency count, numeric}
   \item{inflight}{Frequency count, numeric}
   \item{interest}{Frequency count, numeric}
   \item{member}{Frequency count, numeric}
   \item{delay}{Frequency count, numeric}
   \item{late}{Frequency count, numeric}
   \item{whenev}{Frequency count, numeric}
   \item{cancel}{Frequency count, numeric}
   \item{four}{Frequency count, numeric}
   \item{busi}{Frequency count, numeric}
   \item{disappoint}{Frequency count, numeric}
   \item{meet}{Frequency count, numeric}
   \item{share}{Frequency count, numeric}
   \item{ium}{Frequency count, numeric}
   \item{troubl}{Frequency count, numeric}
   \item{wife}{Frequency count, numeric}
   \item{bring}{Frequency count, numeric}
   \item{code}{Frequency count, numeric}
   \item{bet}{Frequency count, numeric}
   \item{random}{Frequency count, numeric}
   \item{home}{Frequency count, numeric}
   \item{life}{Frequency count, numeric}
   \item{back}{Frequency count, numeric}
   \item{yeah}{Frequency count, numeric}
   \item{point}{Frequency count, numeric}
   \item{recent}{Frequency count, numeric}
   \item{flightl}{Frequency count, numeric}
   \item{jfk}{Frequency count, numeric}
   \item{saturday}{Frequency count, numeric}
   \item{video}{Frequency count, numeric}
   \item{went}{Frequency count, numeric}
   \item{land}{Frequency count, numeric}
   \item{redirect}{Frequency count, numeric}
   \item{btw}{Frequency count, numeric}
   \item{user}{Frequency count, numeric}
   \item{desktop}{Frequency count, numeric}
   \item{mobil}{Frequency count, numeric}
   \item{bin}{Frequency count, numeric}
   \item{class}{Frequency count, numeric}
   \item{passeng}{Frequency count, numeric}
   \item{remov}{Frequency count, numeric}
   \item{scan}{Frequency count, numeric}
   \item{told}{Frequency count, numeric}
   \item{uncomfort}{Frequency count, numeric}
   \item{find}{Frequency count, numeric}
   \item{anyon}{Frequency count, numeric}
   \item{useless}{Frequency count, numeric}
   \item{boy}{Frequency count, numeric}
   \item{thursday}{Frequency count, numeric}
   \item{appoint}{Frequency count, numeric}
   \item{music}{Frequency count, numeric}
   \item{bought}{Frequency count, numeric}
   \item{direct}{Frequency count, numeric}
   \item{fuel}{Frequency count, numeric}
   \item{layov}{Frequency count, numeric}
   \item{unexpect}{Frequency count, numeric}
   \item{vega}{Frequency count, numeric}
   \item{exp}{Frequency count, numeric}
   \item{peopl}{Frequency count, numeric}
   \item{customerservic}{Frequency count, numeric}
   \item{uouo}{Frequency count, numeric}
   \item{hung}{Frequency count, numeric}
   \item{line}{Frequency count, numeric}
   \item{sarcasm}{Frequency count, numeric}
   \item{info}{Frequency count, numeric}
   \item{plain}{Frequency count, numeric}
   \item{text}{Frequency count, numeric}
   \item{version}{Frequency count, numeric}
   \item{look}{Frequency count, numeric}
   \item{weather}{Frequency count, numeric}
   \item{getaway}{Frequency count, numeric}
   \item{lot}{Frequency count, numeric}
   \item{oneway}{Frequency count, numeric}
   \item{uiio}{Frequency count, numeric}
   \item{come}{Frequency count, numeric}
   \item{horribl}{Frequency count, numeric}
   \item{phl}{Frequency count, numeric}
   \item{concern}{Frequency count, numeric}
   \item{tech}{Frequency count, numeric}
   \item{comfort}{Frequency count, numeric}
   \item{easi}{Frequency count, numeric}
   \item{flown}{Frequency count, numeric}
   \item{repres}{Frequency count, numeric}
   \item{beauti}{Frequency count, numeric}
   \item{kick}{Frequency count, numeric}
   \item{rep}{Frequency count, numeric}
   \item{design}{Frequency count, numeric}
   \item{right}{Frequency count, numeric}
   \item{secur}{Frequency count, numeric}
   \item{gate}{Frequency count, numeric}
   \item{kept}{Frequency count, numeric}
   \item{las}{Frequency count, numeric}
   \item{run}{Frequency count, numeric}
   \item{wait}{Frequency count, numeric}
   \item{browser}{Frequency count, numeric}
   \item{folk}{Frequency count, numeric}
   \item{total}{Frequency count, numeric}
   \item{connect}{Frequency count, numeric}
   \item{long}{Frequency count, numeric}
   \item{min}{Frequency count, numeric}
   \item{seem}{Frequency count, numeric}
   \item{francisco}{Frequency count, numeric}
   \item{outstand}{Frequency count, numeric}
   \item{san}{Frequency count, numeric}
   \item{absolut}{Frequency count, numeric}
   \item{delight}{Frequency count, numeric}
   \item{complimentari}{Frequency count, numeric}
   \item{provid}{Frequency count, numeric}
   \item{intern}{Frequency count, numeric}
   \item{that}{Frequency count, numeric}
   \item{complet}{Frequency count, numeric}
   \item{depart}{Frequency count, numeric}
   \item{month}{Frequency count, numeric}
   \item{nonstop}{Frequency count, numeric}
   \item{link}{Frequency count, numeric}
   \item{precheck}{Frequency count, numeric}
   \item{process}{Frequency count, numeric}
   \item{tsa}{Frequency count, numeric}
   \item{hotel}{Frequency count, numeric}
   \item{terribl}{Frequency count, numeric}
   \item{cabo}{Frequency count, numeric}
   \item{assist}{Frequency count, numeric}
   \item{give}{Frequency count, numeric}
   \item{shame}{Frequency count, numeric}
   \item{yesterday}{Frequency count, numeric}
   \item{advis}{Frequency count, numeric}
   \item{dog}{Frequency count, numeric}
   \item{dull}{Frequency count, numeric}
   \item{longer}{Frequency count, numeric}
   \item{pls}{Frequency count, numeric}
   \item{queue}{Frequency count, numeric}
   \item{regist}{Frequency count, numeric}
   \item{redcarpet}{Frequency count, numeric}
   \item{drop}{Frequency count, numeric}
   \item{alway}{Frequency count, numeric}
   \item{buy}{Frequency count, numeric}
   \item{creat}{Frequency count, numeric}
   \item{frustrat}{Frequency count, numeric}
   \item{afford}{Frequency count, numeric}
   \item{buck}{Frequency count, numeric}
   \item{extra}{Frequency count, numeric}
   \item{hire}{Frequency count, numeric}
   \item{lame}{Frequency count, numeric}
   \item{luggag}{Frequency count, numeric}
   \item{might}{Frequency count, numeric}
   \item{cake}{Frequency count, numeric}
   \item{world}{Frequency count, numeric}
   \item{flt}{Frequency count, numeric}
   \item{midnight}{Frequency count, numeric}
   \item{monday}{Frequency count, numeric}
   \item{sooner}{Frequency count, numeric}
   \item{cant}{Frequency count, numeric}
   \item{imagin}{Frequency count, numeric}
   \item{indic}{Frequency count, numeric}
   \item{web}{Frequency count, numeric}
   \item{ran}{Frequency count, numeric}
   \item{staff}{Frequency count, numeric}
   \item{super}{Frequency count, numeric}
   \item{across}{Frequency count, numeric}
   \item{offer}{Frequency count, numeric}
   \item{paid}{Frequency count, numeric}
   \item{premium}{Frequency count, numeric}
   \item{sad}{Frequency count, numeric}
   \item{infant}{Frequency count, numeric}
   \item{possibl}{Frequency count, numeric}
   \item{question}{Frequency count, numeric}
   \item{welcom}{Frequency count, numeric}
   \item{refresh}{Frequency count, numeric}
   \item{damag}{Frequency count, numeric}
   \item{polici}{Frequency count, numeric}
   \item{understand}{Frequency count, numeric}
   \item{destroy}{Frequency count, numeric}
   \item{alon}{Frequency count, numeric}
   \item{board}{Frequency count, numeric}
   \item{kid}{Frequency count, numeric}
   \item{prioriti}{Frequency count, numeric}
   \item{special}{Frequency count, numeric}
   \item{happi}{Frequency count, numeric}
   \item{purchas}{Frequency count, numeric}
   \item{definit}{Frequency count, numeric}
   \item{field}{Frequency count, numeric}
   \item{dfw}{Frequency count, numeric}
   \item{lga}{Frequency count, numeric}
   \item{driver}{Frequency count, numeric}
   \item{friday}{Frequency count, numeric}
   \item{licens}{Frequency count, numeric}
   \item{contact}{Frequency count, numeric}
   \item{error}{Frequency count, numeric}
   \item{keep}{Frequency count, numeric}
   \item{button}{Frequency count, numeric}
   \item{reschedul}{Frequency count, numeric}
   \item{fix}{Frequency count, numeric}
   \item{rang}{Frequency count, numeric}
   \item{got}{Frequency count, numeric}
   \item{tomorrow}{Frequency count, numeric}
   \item{unaccept}{Frequency count, numeric}
   \item{stuck}{Frequency count, numeric}
   \item{includ}{Frequency count, numeric}
   \item{pair}{Frequency count, numeric}
   \item{white}{Frequency count, numeric}
   \item{wine}{Frequency count, numeric}
   \item{agent}{Frequency count, numeric}
   \item{close}{Frequency count, numeric}
   \item{dullesairport}{Frequency count, numeric}
   \item{silver}{Frequency count, numeric}
   \item{austin}{Frequency count, numeric}
   \item{plan}{Frequency count, numeric}
   \item{rout}{Frequency count, numeric}
   \item{central}{Frequency count, numeric}
   \item{normal}{Frequency count, numeric}
   \item{receiv}{Frequency count, numeric}
   \item{repli}{Frequency count, numeric}
   \item{smh}{Frequency count, numeric}
   \item{airport}{Frequency count, numeric}
   \item{checkin}{Frequency count, numeric}
   \item{desk}{Frequency count, numeric}
   \item{logan}{Frequency count, numeric}
   \item{open}{Frequency count, numeric}
   \item{jump}{Frequency count, numeric}
   \item{daili}{Frequency count, numeric}
   \item{luv}{Frequency count, numeric}
   \item{april}{Frequency count, numeric}
   \item{begin}{Frequency count, numeric}
   \item{fuck}{Frequency count, numeric}
   \item{unit}{Frequency count, numeric}
   \item{follow}{Frequency count, numeric}
   \item{apolog}{Frequency count, numeric}
   \item{mani}{Frequency count, numeric}
   \item{respond}{Frequency count, numeric}
   \item{wors}{Frequency count, numeric}
   \item{money}{Frequency count, numeric}
   \item{reimburs}{Frequency count, numeric}
   \item{spend}{Frequency count, numeric}
   \item{strand}{Frequency count, numeric}
   \item{choic}{Frequency count, numeric}
   \item{sjc}{Frequency count, numeric}
   \item{southwest}{Frequency count, numeric}
   \item{claim}{Frequency count, numeric}
   \item{confirm}{Frequency count, numeric}
   \item{day}{Frequency count, numeric}
   \item{inquir}{Frequency count, numeric}
   \item{bos}{Frequency count, numeric}
   \item{notif}{Frequency count, numeric}
   \item{rebook}{Frequency count, numeric}
   \item{met}{Frequency count, numeric}
   \item{gold}{Frequency count, numeric}
   \item{cmh}{Frequency count, numeric}
   \item{see}{Frequency count, numeric}
   \item{worst}{Frequency count, numeric}
   \item{stuff}{Frequency count, numeric}
   \item{wrong}{Frequency count, numeric}
   \item{head}{Frequency count, numeric}
   \item{issu}{Frequency count, numeric}
   \item{item}{Frequency count, numeric}
   \item{americanairlin}{Frequency count, numeric}
   \item{lose}{Frequency count, numeric}
   \item{suck}{Frequency count, numeric}
   \item{dmd}{Frequency count, numeric}
   \item{drive}{Frequency count, numeric}
   \item{flightlat}{Frequency count, numeric}
   \item{mon}{Frequency count, numeric}
   \item{notic}{Frequency count, numeric}
   \item{boston}{Frequency count, numeric}
   \item{abil}{Frequency count, numeric}
   \item{expect}{Frequency count, numeric}
   \item{stink}{Frequency count, numeric}
   \item{danc}{Frequency count, numeric}
   \item{high}{Frequency count, numeric}
   \item{school}{Frequency count, numeric}
   \item{appar}{Frequency count, numeric}
   \item{faster}{Frequency count, numeric}
   \item{quick}{Frequency count, numeric}
   \item{nervous}{Frequency count, numeric}
   \item{sent}{Frequency count, numeric}
   \item{pictur}{Frequency count, numeric}
   \item{wing}{Frequency count, numeric}
   \item{crisi}{Frequency count, numeric}
   \item{pressur}{Frequency count, numeric}
   \item{texa}{Frequency count, numeric}
   \item{job}{Frequency count, numeric}
   \item{pilot}{Frequency count, numeric}
   \item{snow}{Frequency count, numeric}
   \item{afternoon}{Frequency count, numeric}
   \item{area}{Frequency count, numeric}
   \item{york}{Frequency count, numeric}
   \item{mention}{Frequency count, numeric}
   \item{reclin}{Frequency count, numeric}
   \item{row}{Frequency count, numeric}
   \item{drink}{Frequency count, numeric}
   \item{handl}{Frequency count, numeric}
   \item{mechan}{Frequency count, numeric}
   \item{scare}{Frequency count, numeric}
   \item{kill}{Frequency count, numeric}
   \item{twitter}{Frequency count, numeric}
   \item{charg}{Frequency count, numeric}
   \item{refund}{Frequency count, numeric}
   \item{refus}{Frequency count, numeric}
   \item{access}{Frequency count, numeric}
   \item{canut}{Frequency count, numeric}
   \item{ipad}{Frequency count, numeric}
   \item{iphon}{Frequency count, numeric}
   \item{passbook}{Frequency count, numeric}
   \item{click}{Frequency count, numeric}
   \item{broken}{Frequency count, numeric}
   \item{forward}{Frequency count, numeric}
   \item{sec}{Frequency count, numeric}
   \item{cost}{Frequency count, numeric}
   \item{ugh}{Frequency count, numeric}
   \item{along}{Frequency count, numeric}
   \item{pdx}{Frequency count, numeric}
   \item{put}{Frequency count, numeric}
   \item{gonna}{Frequency count, numeric}
   \item{bug}{Frequency count, numeric}
   \item{itus}{Frequency count, numeric}
   \item{report}{Frequency count, numeric}
   \item{bit}{Frequency count, numeric}
   \item{center}{Frequency count, numeric}
   \item{attend}{Frequency count, numeric}
   \item{password}{Frequency count, numeric}
   \item{reset}{Frequency count, numeric}
   \item{base}{Frequency count, numeric}
   \item{disrupt}{Frequency count, numeric}
   \item{someth}{Frequency count, numeric}
   \item{departur}{Frequency count, numeric}
   \item{palm}{Frequency count, numeric}
   \item{spring}{Frequency count, numeric}
   \item{fun}{Frequency count, numeric}
   \item{sound}{Frequency count, numeric}
   \item{updat}{Frequency count, numeric}
   \item{count}{Frequency count, numeric}
   \item{somehow}{Frequency count, numeric}
   \item{continu}{Frequency count, numeric}
   \item{sort}{Frequency count, numeric}
   \item{sweet}{Frequency count, numeric}
   \item{coffe}{Frequency count, numeric}
   \item{ice}{Frequency count, numeric}
   \item{assault}{Frequency count, numeric}
   \item{file}{Frequency count, numeric}
   \item{trust}{Frequency count, numeric}
   \item{destin}{Frequency count, numeric}
   \item{sever}{Frequency count, numeric}
   \item{wheel}{Frequency count, numeric}
   \item{itinerari}{Frequency count, numeric}
   \item{abq}{Frequency count, numeric}
   \item{becom}{Frequency count, numeric}
   \item{event}{Frequency count, numeric}
   \item{girl}{Frequency count, numeric}
   \item{inspir}{Frequency count, numeric}
   \item{march}{Frequency count, numeric}
   \item{entir}{Frequency count, numeric}
   \item{improv}{Frequency count, numeric}
   \item{request}{Frequency count, numeric}
   \item{treat}{Frequency count, numeric}
   \item{window}{Frequency count, numeric}
   \item{grand}{Frequency count, numeric}
   \item{exhaust}{Frequency count, numeric}
   \item{flier}{Frequency count, numeric}
   \item{frequent}{Frequency count, numeric}
   \item{inconveni}{Frequency count, numeric}
   \item{purpos}{Frequency count, numeric}
   \item{wtf}{Frequency count, numeric}
   \item{full}{Frequency count, numeric}
   \item{boo}{Frequency count, numeric}
   \item{place}{Frequency count, numeric}
   \item{result}{Frequency count, numeric}
   \item{usa}{Frequency count, numeric}
   \item{instead}{Frequency count, numeric}
   \item{pillow}{Frequency count, numeric}
   \item{race}{Frequency count, numeric}
   \item{advantag}{Frequency count, numeric}
   \item{fight}{Frequency count, numeric}
   \item{salt}{Frequency count, numeric}
   \item{behind}{Frequency count, numeric}
   \item{cabin}{Frequency count, numeric}
   \item{fav}{Frequency count, numeric}
   \item{hot}{Frequency count, numeric}
   \item{pari}{Frequency count, numeric}
   \item{uio}{Frequency count, numeric}
   \item{surgeri}{Frequency count, numeric}
   \item{tell}{Frequency count, numeric}
   \item{suitcas}{Frequency count, numeric}
   \item{past}{Frequency count, numeric}
   \item{order}{Frequency count, numeric}
   \item{regul}{Frequency count, numeric}
   \item{atlant}{Frequency count, numeric}
   \item{benefit}{Frequency count, numeric}
   \item{virginatlant}{Frequency count, numeric}
   \item{came}{Frequency count, numeric}
   \item{husband}{Frequency count, numeric}
   \item{screen}{Frequency count, numeric}
   \item{bare}{Frequency count, numeric}
   \item{light}{Frequency count, numeric}
   \item{turbul}{Frequency count, numeric}
   \item{profit}{Frequency count, numeric}
   \item{funni}{Frequency count, numeric}
   \item{stori}{Frequency count, numeric}
   \item{forecast}{Frequency count, numeric}
   \item{nytim}{Frequency count, numeric}
   \item{applic}{Frequency count, numeric}
   \item{flylaxairport}{Frequency count, numeric}
   \item{posit}{Frequency count, numeric}
   \item{wonder}{Frequency count, numeric}
   \item{despit}{Frequency count, numeric}
   \item{revenu}{Frequency count, numeric}
   \item{competit}{Frequency count, numeric}
   \item{increas}{Frequency count, numeric}
   \item{sos}{Frequency count, numeric}
   \item{degre}{Frequency count, numeric}
   \item{wind}{Frequency count, numeric}
   \item{safeti}{Frequency count, numeric}
   \item{wanna}{Frequency count, numeric}
   \item{hope}{Frequency count, numeric}
   \item{newark}{Frequency count, numeric}
   \item{inappropri}{Frequency count, numeric}
   \item{probabl}{Frequency count, numeric}
   \item{save}{Frequency count, numeric}
   \item{standbi}{Frequency count, numeric}
   \item{comp}{Frequency count, numeric}
   \item{brought}{Frequency count, numeric}
   \item{readi}{Frequency count, numeric}
   \item{manag}{Frequency count, numeric}
   \item{battl}{Frequency count, numeric}
   \item{old}{Frequency count, numeric}
   \item{treatment}{Frequency count, numeric}
   \item{uou}{Frequency count, numeric}
   \item{urgent}{Frequency count, numeric}
   \item{half}{Frequency count, numeric}
   \item{less}{Frequency count, numeric}
   \item{price}{Frequency count, numeric}
   \item{round}{Frequency count, numeric}
   \item{stun}{Frequency count, numeric}
   \item{dtw}{Frequency count, numeric}
   \item{sold}{Frequency count, numeric}
   \item{promo}{Frequency count, numeric}
   \item{made}{Frequency count, numeric}
   \item{paperwork}{Frequency count, numeric}
   \item{promis}{Frequency count, numeric}
   \item{resolut}{Frequency count, numeric}
   \item{perform}{Frequency count, numeric}
   \item{care}{Frequency count, numeric}
   \item{challeng}{Frequency count, numeric}
   \item{mileag}{Frequency count, numeric}
   \item{ball}{Frequency count, numeric}
   \item{americanair}{Frequency count, numeric}
   \item{gma}{Frequency count, numeric}
   \item{glad}{Frequency count, numeric}
   \item{took}{Frequency count, numeric}
   \item{earn}{Frequency count, numeric}
   \item{cheer}{Frequency count, numeric}
   \item{coupl}{Frequency count, numeric}
   \item{messag}{Frequency count, numeric}
   \item{easili}{Frequency count, numeric}
   \item{smooth}{Frequency count, numeric}
   \item{empti}{Frequency count, numeric}
   \item{forc}{Frequency count, numeric}
   \item{small}{Frequency count, numeric}
   \item{space}{Frequency count, numeric}
   \item{slow}{Frequency count, numeric}
   \item{loyal}{Frequency count, numeric}
   \item{solut}{Frequency count, numeric}
   \item{spoke}{Frequency count, numeric}
   \item{jacket}{Frequency count, numeric}
   \item{convers}{Frequency count, numeric}
   \item{method}{Frequency count, numeric}
   \item{partner}{Frequency count, numeric}
   \item{payment}{Frequency count, numeric}
   \item{rate}{Frequency count, numeric}
   \item{visa}{Frequency count, numeric}
   \item{promot}{Frequency count, numeric}
   \item{recommend}{Frequency count, numeric}
   \item{aus}{Frequency count, numeric}
   \item{turn}{Frequency count, numeric}
   \item{cours}{Frequency count, numeric}
   \item{famili}{Frequency count, numeric}
   \item{receipt}{Frequency count, numeric}
   \item{badg}{Frequency count, numeric}
   \item{superior}{Frequency count, numeric}
   \item{standard}{Frequency count, numeric}
   \item{pick}{Frequency count, numeric}
   \item{tweet}{Frequency count, numeric}
   \item{car}{Frequency count, numeric}
   \item{rental}{Frequency count, numeric}
   \item{child}{Frequency count, numeric}
   \item{cust}{Frequency count, numeric}
   \item{lap}{Frequency count, numeric}
   \item{chanc}{Frequency count, numeric}
   \item{gift}{Frequency count, numeric}
   \item{bank}{Frequency count, numeric}
   \item{constant}{Frequency count, numeric}
   \item{joke}{Frequency count, numeric}
   \item{also}{Frequency count, numeric}
   \item{bday}{Frequency count, numeric}
   \item{eastern}{Frequency count, numeric}
   \item{ella}{Frequency count, numeric}
   \item{grandma}{Frequency count, numeric}
   \item{greet}{Frequency count, numeric}
   \item{stewardess}{Frequency count, numeric}
   \item{tix}{Frequency count, numeric}
   \item{transact}{Frequency count, numeric}
   \item{moment}{Frequency count, numeric}
   \item{send}{Frequency count, numeric}
   \item{havenut}{Frequency count, numeric}
   \item{submit}{Frequency count, numeric}
   \item{fat}{Frequency count, numeric}
   \item{ceo}{Frequency count, numeric}
   \item{low}{Frequency count, numeric}
   \item{model}{Frequency count, numeric}
   \item{brilliant}{Frequency count, numeric}
   \item{omg}{Frequency count, numeric}
   \item{thought}{Frequency count, numeric}
   \item{comput}{Frequency count, numeric}
   \item{pull}{Frequency count, numeric}
   \item{incid}{Frequency count, numeric}
   \item{laptop}{Frequency count, numeric}
   \item{guess}{Frequency count, numeric}
   \item{road}{Frequency count, numeric}
   \item{announc}{Frequency count, numeric}
   \item{dollar}{Frequency count, numeric}
   \item{deserv}{Frequency count, numeric}
   \item{address}{Frequency count, numeric}
   \item{bounc}{Frequency count, numeric}
   \item{exist}{Frequency count, numeric}
   \item{fyi}{Frequency count, numeric}
   \item{requir}{Frequency count, numeric}
   \item{svc}{Frequency count, numeric}
   \item{toward}{Frequency count, numeric}
   \item{app}{Frequency count, numeric}
   \item{develop}{Frequency count, numeric}
   \item{favorit}{Frequency count, numeric}
   \item{channel}{Frequency count, numeric}
   \item{crap}{Frequency count, numeric}
   \item{set}{Frequency count, numeric}
   \item{integr}{Frequency count, numeric}
   \item{mexico}{Frequency count, numeric}
   \item{paper}{Frequency count, numeric}
   \item{wast}{Frequency count, numeric}
   \item{accept}{Frequency count, numeric}
   \item{system}{Frequency count, numeric}
   \item{black}{Frequency count, numeric}
   \item{tea}{Frequency count, numeric}
   \item{spread}{Frequency count, numeric}
   \item{learn}{Frequency count, numeric}
   \item{experienc}{Frequency count, numeric}
   \item{incred}{Frequency count, numeric}
   \item{philadelphia}{Frequency count, numeric}
   \item{houston}{Frequency count, numeric}
   \item{final}{Frequency count, numeric}
   \item{avgeek}{Frequency count, numeric}
   \item{forget}{Frequency count, numeric}
   \item{futur}{Frequency count, numeric}
   \item{jack}{Frequency count, numeric}
   \item{kinda}{Frequency count, numeric}
   \item{diego}{Frequency count, numeric}
   \item{malfunct}{Frequency count, numeric}
   \item{poor}{Frequency count, numeric}
   \item{thankyou}{Frequency count, numeric}
   \item{jose}{Frequency count, numeric}
   \item{compet}{Frequency count, numeric}
   \item{excel}{Frequency count, numeric}
   \item{portland}{Frequency count, numeric}
   \item{addit}{Frequency count, numeric}
   \item{thur}{Frequency count, numeric}
   \item{uau}{Frequency count, numeric}
   \item{carryon}{Frequency count, numeric}
   \item{lock}{Frequency count, numeric}
   \item{group}{Frequency count, numeric}
   \item{room}{Frequency count, numeric}
   \item{nightmar}{Frequency count, numeric}
   \item{classi}{Frequency count, numeric}
   \item{strong}{Frequency count, numeric}
   \item{sure}{Frequency count, numeric}
   \item{shit}{Frequency count, numeric}
   \item{bump}{Frequency count, numeric}
   \item{luckili}{Frequency count, numeric}
   \item{rough}{Frequency count, numeric}
   \item{perfect}{Frequency count, numeric}
   \item{rais}{Frequency count, numeric}
   \item{stay}{Frequency count, numeric}
   \item{sun}{Frequency count, numeric}
   \item{man}{Frequency count, numeric}
   \item{wifi}{Frequency count, numeric}
   \item{donut}{Frequency count, numeric}
   \item{iull}{Frequency count, numeric}
   \item{saw}{Frequency count, numeric}
   \item{dad}{Frequency count, numeric}
   \item{part}{Frequency count, numeric}
   \item{whole}{Frequency count, numeric}
   \item{europ}{Frequency count, numeric}
   \item{short}{Frequency count, numeric}
   \item{abc}{Frequency count, numeric}
   \item{overhead}{Frequency count, numeric}
   \item{person}{Frequency count, numeric}
   \item{ask}{Frequency count, numeric}
   \item{luck}{Frequency count, numeric}
   \item{accommod}{Frequency count, numeric}
   \item{dear}{Frequency count, numeric}
   \item{lol}{Frequency count, numeric}
   \item{actual}{Frequency count, numeric}
   \item{other}{Frequency count, numeric}
   \item{serv}{Frequency count, numeric}
   \item{affect}{Frequency count, numeric}
   \item{shock}{Frequency count, numeric}
   \item{storm}{Frequency count, numeric}
   \item{umm}{Frequency count, numeric}
   \item{advanc}{Frequency count, numeric}
   \item{cut}{Frequency count, numeric}
   \item{philli}{Frequency count, numeric}
   \item{main}{Frequency count, numeric}
   \item{penalti}{Frequency count, numeric}
   \item{industri}{Frequency count, numeric}
   \item{bonus}{Frequency count, numeric}
   \item{regard}{Frequency count, numeric}
   \item{movi}{Frequency count, numeric}
   \item{resolv}{Frequency count, numeric}
   \item{word}{Frequency count, numeric}
   \item{deic}{Frequency count, numeric}
   \item{lack}{Frequency count, numeric}
   \item{common}{Frequency count, numeric}
   \item{fill}{Frequency count, numeric}
   \item{express}{Frequency count, numeric}
   \item{hop}{Frequency count, numeric}
   \item{huge}{Frequency count, numeric}
   \item{popular}{Frequency count, numeric}
   \item{orlando}{Frequency count, numeric}
   \item{accid}{Frequency count, numeric}
   \item{tough}{Frequency count, numeric}
   \item{altern}{Frequency count, numeric}
   \item{hook}{Frequency count, numeric}
   \item{prevent}{Frequency count, numeric}
   \item{domest}{Frequency count, numeric}
   \item{law}{Frequency count, numeric}
   \item{auto}{Frequency count, numeric}
   \item{economi}{Frequency count, numeric}
   \item{logic}{Frequency count, numeric}
   \item{transfer}{Frequency count, numeric}
   \item{shoe}{Frequency count, numeric}
   \item{case}{Frequency count, numeric}
   \item{david}{Frequency count, numeric}
   \item{cousin}{Frequency count, numeric}
   \item{januari}{Frequency count, numeric}
   \item{voucher}{Frequency count, numeric}
   \item{mail}{Frequency count, numeric}
   \item{compani}{Frequency count, numeric}
   \item{execut}{Frequency count, numeric}
   \item{manner}{Frequency count, numeric}
   \item{mayb}{Frequency count, numeric}
   \item{chase}{Frequency count, numeric}
   \item{pass}{Frequency count, numeric}
   \item{son}{Frequency count, numeric}
   \item{weak}{Frequency count, numeric}
   \item{didnt}{Frequency count, numeric}
   \item{given}{Frequency count, numeric}
   \item{jan}{Frequency count, numeric}
   \item{miami}{Frequency count, numeric}
   \item{septemb}{Frequency count, numeric}
   \item{explain}{Frequency count, numeric}
   \item{talk}{Frequency count, numeric}
   \item{almost}{Frequency count, numeric}
   \item{daughter}{Frequency count, numeric}
   \item{ppl}{Frequency count, numeric}
   \item{flightr}{Frequency count, numeric}
   \item{repeat}{Frequency count, numeric}
   \item{anymor}{Frequency count, numeric}
   \item{platinum}{Frequency count, numeric}
   \item{termin}{Frequency count, numeric}
   \item{badcustomerservic}{Frequency count, numeric}
   \item{instruct}{Frequency count, numeric}
   \item{rude}{Frequency count, numeric}
   \item{overnight}{Frequency count, numeric}
   \item{client}{Frequency count, numeric}
   \item{iah}{Frequency count, numeric}
   \item{taken}{Frequency count, numeric}
   \item{accord}{Frequency count, numeric}
   \item{illeg}{Frequency count, numeric}
   \item{detail}{Frequency count, numeric}
   \item{dmed}{Frequency count, numeric}
   \item{charlott}{Frequency count, numeric}
   \item{ignor}{Frequency count, numeric}
   \item{touch}{Frequency count, numeric}
   \item{form}{Frequency count, numeric}
   \item{anywher}{Frequency count, numeric}
   \item{upon}{Frequency count, numeric}
   \item{ground}{Frequency count, numeric}
   \item{cessna}{Frequency count, numeric}
   \item{els}{Frequency count, numeric}
   \item{appreci}{Frequency count, numeric}
   \item{tarmac}{Frequency count, numeric}
   \item{outsid}{Frequency count, numeric}
   \item{break}{Frequency count, numeric}
   \item{guitar}{Frequency count, numeric}
   \item{truli}{Frequency count, numeric}
   \item{besid}{Frequency count, numeric}
   \item{third}{Frequency count, numeric}
   \item{disabl}{Frequency count, numeric}
   \item{search}{Frequency count, numeric}
   \item{enough}{Frequency count, numeric}
   \item{rule}{Frequency count, numeric}
   \item{denver}{Frequency count, numeric}
   \item{mistak}{Frequency count, numeric}
   \item{train}{Frequency count, numeric}
   \item{ord}{Frequency count, numeric}
   \item{avoid}{Frequency count, numeric}
   \item{found}{Frequency count, numeric}
   \item{supervisor}{Frequency count, numeric}
   \item{pre}{Frequency count, numeric}
   \item{preboard}{Frequency count, numeric}
   \item{dept}{Frequency count, numeric}
   \item{okc}{Frequency count, numeric}
   \item{thru}{Frequency count, numeric}
   \item{employe}{Frequency count, numeric}
   \item{equip}{Frequency count, numeric}
   \item{medic}{Frequency count, numeric}
   \item{upset}{Frequency count, numeric}
   \item{consider}{Frequency count, numeric}
   \item{intl}{Frequency count, numeric}
   \item{thnx}{Frequency count, numeric}
   \item{activ}{Frequency count, numeric}
   \item{emerg}{Frequency count, numeric}
   \item{militari}{Frequency count, numeric}
   \item{mom}{Frequency count, numeric}
   \item{wont}{Frequency count, numeric}
   \item{mile}{Frequency count, numeric}
   \item{mileageplus}{Frequency count, numeric}
   \item{complain}{Frequency count, numeric}
   \item{throw}{Frequency count, numeric}
   \item{unfortun}{Frequency count, numeric}
   \item{american}{Frequency count, numeric}
   \item{ahead}{Frequency count, numeric}
   \item{bird}{Frequency count, numeric}
   \item{coach}{Frequency count, numeric}
   \item{control}{Frequency count, numeric}
   \item{rather}{Frequency count, numeric}
   \item{sincer}{Frequency count, numeric}
   \item{bother}{Frequency count, numeric}
   \item{delta}{Frequency count, numeric}
   \item{unhelp}{Frequency count, numeric}
   \item{attempt}{Frequency count, numeric}
   \item{buddi}{Frequency count, numeric}
   \item{neglig}{Frequency count, numeric}
   \item{believ}{Frequency count, numeric}
   \item{fll}{Frequency count, numeric}
   \item{disappear}{Frequency count, numeric}
   \item{stand}{Frequency count, numeric}
   \item{record}{Frequency count, numeric}
   \item{nut}{Frequency count, numeric}
   \item{guid}{Frequency count, numeric}
   \item{read}{Frequency count, numeric}
   \item{bogota}{Frequency count, numeric}
   \item{everyon}{Frequency count, numeric}
   \item{inform}{Frequency count, numeric}
   \item{outbound}{Frequency count, numeric}
   \item{aircraft}{Frequency count, numeric}
   \item{switch}{Frequency count, numeric}
   \item{crash}{Frequency count, numeric}
   \item{knew}{Frequency count, numeric}
   \item{negoti}{Frequency count, numeric}
   \item{winter}{Frequency count, numeric}
   \item{coat}{Frequency count, numeric}
   \item{scam}{Frequency count, numeric}
   \item{reach}{Frequency count, numeric}
   \item{elsewher}{Frequency count, numeric}
   \item{honor}{Frequency count, numeric}
   \item{hear}{Frequency count, numeric}
   \item{quit}{Frequency count, numeric}
   \item{pathet}{Frequency count, numeric}
   \item{boot}{Frequency count, numeric}
   \item{unavail}{Frequency count, numeric}
   \item{minor}{Frequency count, numeric}
   \item{babi}{Frequency count, numeric}
   \item{mine}{Frequency count, numeric}
   \item{huh}{Frequency count, numeric}
   \item{goe}{Frequency count, numeric}
   \item{obvious}{Frequency count, numeric}
   \item{simpl}{Frequency count, numeric}
   \item{disrespect}{Frequency count, numeric}
   \item{neglect}{Frequency count, numeric}
   \item{enjoy}{Frequency count, numeric}
   \item{finger}{Frequency count, numeric}
   \item{zone}{Frequency count, numeric}
   \item{date}{Frequency count, numeric}
   \item{excus}{Frequency count, numeric}
   \item{zero}{Frequency count, numeric}
   \item{decemb}{Frequency count, numeric}
   \item{octob}{Frequency count, numeric}
   \item{sna}{Frequency count, numeric}
   \item{weekend}{Frequency count, numeric}
   \item{flightd}{Frequency count, numeric}
   \item{local}{Frequency count, numeric}
   \item{faith}{Frequency count, numeric}
   \item{ski}{Frequency count, numeric}
   \item{held}{Frequency count, numeric}
   \item{coast}{Frequency count, numeric}
   \item{tourist}{Frequency count, numeric}
   \item{listen}{Frequency count, numeric}
   \item{whoever}{Frequency count, numeric}
   \item{charact}{Frequency count, numeric}
   \item{limit}{Frequency count, numeric}
   \item{appear}{Frequency count, numeric}
   \item{doesnt}{Frequency count, numeric}
   \item{pop}{Frequency count, numeric}
   \item{coupon}{Frequency count, numeric}
   \item{ridicul}{Frequency count, numeric}
   \item{airfar}{Frequency count, numeric}
   \item{cheaper}{Frequency count, numeric}
   \item{heathrow}{Frequency count, numeric}
   \item{properti}{Frequency count, numeric}
   \item{complaint}{Frequency count, numeric}
   \item{ref}{Frequency count, numeric}
   \item{unhappi}{Frequency count, numeric}
   \item{jet}{Frequency count, numeric}
   \item{privat}{Frequency count, numeric}
   \item{fund}{Frequency count, numeric}
   \item{atl}{Frequency count, numeric}
   \item{matter}{Frequency count, numeric}
   \item{yep}{Frequency count, numeric}
   \item{loui}{Frequency count, numeric}
   \item{difficult}{Frequency count, numeric}
   \item{prepar}{Frequency count, numeric}
   \item{confid}{Frequency count, numeric}
   \item{seek}{Frequency count, numeric}
   \item{bot}{Frequency count, numeric}
   \item{worth}{Frequency count, numeric}
   \item{california}{Frequency count, numeric}
   \item{conf}{Frequency count, numeric}
   \item{communic}{Frequency count, numeric}
   \item{far}{Frequency count, numeric}
   \item{horrid}{Frequency count, numeric}
   \item{note}{Frequency count, numeric}
   \item{top}{Frequency count, numeric}
   \item{unitedairlin}{Frequency count, numeric}
   \item{kudo}{Frequency count, numeric}
   \item{usual}{Frequency count, numeric}
   \item{advisori}{Frequency count, numeric}
   \item{quot}{Frequency count, numeric}
   \item{heat}{Frequency count, numeric}
   \item{rememb}{Frequency count, numeric}
   \item{captain}{Frequency count, numeric}
   \item{extrem}{Frequency count, numeric}
   \item{effici}{Frequency count, numeric}
   \item{real}{Frequency count, numeric}
   \item{henc}{Frequency count, numeric}
   \item{cloth}{Frequency count, numeric}
   \item{curious}{Frequency count, numeric}
   \item{conveni}{Frequency count, numeric}
   \item{washington}{Frequency count, numeric}
   \item{chicago}{Frequency count, numeric}
   \item{china}{Frequency count, numeric}
   \item{leg}{Frequency count, numeric}
   \item{loung}{Frequency count, numeric}
   \item{dms}{Frequency count, numeric}
   \item{owe}{Frequency count, numeric}
   \item{okay}{Frequency count, numeric}
   \item{dead}{Frequency count, numeric}
   \item{proper}{Frequency count, numeric}
   \item{forgot}{Frequency count, numeric}
   \item{onto}{Frequency count, numeric}
   \item{atroci}{Frequency count, numeric}
   \item{imposs}{Frequency count, numeric}
   \item{esp}{Frequency count, numeric}
   \item{lhr}{Frequency count, numeric}
   \item{doubt}{Frequency count, numeric}
   \item{vacat}{Frequency count, numeric}
   \item{compliment}{Frequency count, numeric}
   \item{assur}{Frequency count, numeric}
   \item{nope}{Frequency count, numeric}
   \item{sleep}{Frequency count, numeric}
   \item{compens}{Frequency count, numeric}
   \item{origin}{Frequency count, numeric}
   \item{ruin}{Frequency count, numeric}
   \item{defin}{Frequency count, numeric}
   \item{inbound}{Frequency count, numeric}
   \item{valid}{Frequency count, numeric}
   \item{deliv}{Frequency count, numeric}
   \item{deliveri}{Frequency count, numeric}
   \item{hostag}{Frequency count, numeric}
   \item{lbs}{Frequency count, numeric}
   \item{machin}{Frequency count, numeric}
   \item{write}{Frequency count, numeric}
   \item{blame}{Frequency count, numeric}
   \item{previous}{Frequency count, numeric}
   \item{tel}{Frequency count, numeric}
   \item{aviv}{Frequency count, numeric}
   \item{uiunit}{Frequency count, numeric}
   \item{feedback}{Frequency count, numeric}
   \item{taxi}{Frequency count, numeric}
   \item{action}{Frequency count, numeric}
   \item{solv}{Frequency count, numeric}
   \item{tire}{Frequency count, numeric}
   \item{comment}{Frequency count, numeric}
   \item{clt}{Frequency count, numeric}
   \item{idea}{Frequency count, numeric}
   \item{neveragain}{Frequency count, numeric}
   \item{within}{Frequency count, numeric}
   \item{cancun}{Frequency count, numeric}
   \item{continent}{Frequency count, numeric}
   \item{equal}{Frequency count, numeric}
   \item{deep}{Frequency count, numeric}
   \item{figur}{Frequency count, numeric}
   \item{fran}{Frequency count, numeric}
   \item{pack}{Frequency count, numeric}
   \item{pax}{Frequency count, numeric}
   \item{screw}{Frequency count, numeric}
   \item{hou}{Frequency count, numeric}
   \item{pos}{Frequency count, numeric}
   \item{wake}{Frequency count, numeric}
   \item{counter}{Frequency count, numeric}
   \item{nobodi}{Frequency count, numeric}
   \item{flyohar}{Frequency count, numeric}
   \item{regret}{Frequency count, numeric}
   \item{annoy}{Frequency count, numeric}
   \item{door}{Frequency count, numeric}
   \item{premier}{Frequency count, numeric}
   \item{bullshit}{Frequency count, numeric}
   \item{phoenix}{Frequency count, numeric}
   \item{yea}{Frequency count, numeric}
   \item{encount}{Frequency count, numeric}
   \item{stellar}{Frequency count, numeric}
   \item{front}{Frequency count, numeric}
   \item{pocket}{Frequency count, numeric}
   \item{rip}{Frequency count, numeric}
   \item{absurd}{Frequency count, numeric}
   \item{fact}{Frequency count, numeric}
   \item{walk}{Frequency count, numeric}
   \item{happili}{Frequency count, numeric}
   \item{public}{Frequency count, numeric}
   \item{stranger}{Frequency count, numeric}
   \item{nah}{Frequency count, numeric}
   \item{split}{Frequency count, numeric}
   \item{wks}{Frequency count, numeric}
   \item{bwi}{Frequency count, numeric}
   \item{appropri}{Frequency count, numeric}
   \item{review}{Frequency count, numeric}
   \item{overbook}{Frequency count, numeric}
   \item{unprofession}{Frequency count, numeric}
   \item{consid}{Frequency count, numeric}
   \item{corp}{Frequency count, numeric}
   \item{unabl}{Frequency count, numeric}
   \item{flightat}{Frequency count, numeric}
   \item{courtesi}{Frequency count, numeric}
   \item{diff}{Frequency count, numeric}
   \item{tag}{Frequency count, numeric}
   \item{certif}{Frequency count, numeric}
   \item{bar}{Frequency count, numeric}
   \item{appl}{Frequency count, numeric}
   \item{bunch}{Frequency count, numeric}
   \item{adopt}{Frequency count, numeric}
   \item{outlet}{Frequency count, numeric}
   \item{incompet}{Frequency count, numeric}
   \item{finish}{Frequency count, numeric}
   \item{exact}{Frequency count, numeric}
   \item{carousel}{Frequency count, numeric}
   \item{easier}{Frequency count, numeric}
   \item{monitor}{Frequency count, numeric}
   \item{friendlyski}{Frequency count, numeric}
   \item{spare}{Frequency count, numeric}
   \item{unless}{Frequency count, numeric}
   \item{disast}{Frequency count, numeric}
   \item{period}{Frequency count, numeric}
   \item{apart}{Frequency count, numeric}
   \item{fall}{Frequency count, numeric}
   \item{fleet}{Frequency count, numeric}
   \item{notcool}{Frequency count, numeric}
   \item{balanc}{Frequency count, numeric}
   \item{impact}{Frequency count, numeric}
   \item{major}{Frequency count, numeric}
   \item{refuel}{Frequency count, numeric}
   \item{hundr}{Frequency count, numeric}
   \item{quiet}{Frequency count, numeric}
   \item{rectifi}{Frequency count, numeric}
   \item{situat}{Frequency count, numeric}
   \item{spent}{Frequency count, numeric}
   \item{clue}{Frequency count, numeric}
   \item{liter}{Frequency count, numeric}
   \item{list}{Frequency count, numeric}
   \item{dia}{Frequency count, numeric}
   \item{remind}{Frequency count, numeric}
   \item{devic}{Frequency count, numeric}
   \item{insid}{Frequency count, numeric}
   \item{failur}{Frequency count, numeric}
   \item{assum}{Frequency count, numeric}
   \item{effort}{Frequency count, numeric}
   \item{finest}{Frequency count, numeric}
   \item{confus}{Frequency count, numeric}
   \item{airplan}{Frequency count, numeric}
   \item{project}{Frequency count, numeric}
   \item{softwar}{Frequency count, numeric}
   \item{ogg}{Frequency count, numeric}
   \item{display}{Frequency count, numeric}
   \item{howev}{Frequency count, numeric}
   \item{closest}{Frequency count, numeric}
   \item{clear}{Frequency count, numeric}
   \item{refer}{Frequency count, numeric}
   \item{discuss}{Frequency count, numeric}
   \item{shot}{Frequency count, numeric}
   \item{pacif}{Frequency count, numeric}
   \item{spot}{Frequency count, numeric}
   \item{generic}{Frequency count, numeric}
   \item{unnecessari}{Frequency count, numeric}
   \item{carseat}{Frequency count, numeric}
   \item{decis}{Frequency count, numeric}
   \item{kpu}{Frequency count, numeric}
   \item{kit}{Frequency count, numeric}
   \item{pjs}{Frequency count, numeric}
   \item{ensur}{Frequency count, numeric}
   \item{flat}{Frequency count, numeric}
   \item{lie}{Frequency count, numeric}
   \item{stl}{Frequency count, numeric}
   \item{membership}{Frequency count, numeric}
   \item{catch}{Frequency count, numeric}
   \item{karen}{Frequency count, numeric}
   \item{chicken}{Frequency count, numeric}
   \item{fair}{Frequency count, numeric}
   \item{den}{Frequency count, numeric}
   \item{ohar}{Frequency count, numeric}
   \item{reinstat}{Frequency count, numeric}
   \item{offic}{Frequency count, numeric}
   \item{attach}{Frequency count, numeric}
   \item{photo}{Frequency count, numeric}
   \item{damn}{Frequency count, numeric}
   \item{canada}{Frequency count, numeric}
   \item{yvr}{Frequency count, numeric}
   \item{divert}{Frequency count, numeric}
   \item{msi}{Frequency count, numeric}
   \item{canadian}{Frequency count, numeric}
   \item{male}{Frequency count, numeric}
   \item{pic}{Frequency count, numeric}
   \item{threaten}{Frequency count, numeric}
   \item{yell}{Frequency count, numeric}
   \item{tho}{Frequency count, numeric}
   \item{frm}{Frequency count, numeric}
   \item{reflight}{Frequency count, numeric}
   \item{sea}{Frequency count, numeric}
   \item{connector}{Frequency count, numeric}
   \item{gross}{Frequency count, numeric}
   \item{exampl}{Frequency count, numeric}
   \item{statement}{Frequency count, numeric}
   \item{otherwis}{Frequency count, numeric}
   \item{stood}{Frequency count, numeric}
   \item{inadequ}{Frequency count, numeric}
   \item{unfriendlyski}{Frequency count, numeric}
   \item{seen}{Frequency count, numeric}
   \item{singl}{Frequency count, numeric}
   \item{locat}{Frequency count, numeric}
   \item{boyfriend}{Frequency count, numeric}
   \item{gotta}{Frequency count, numeric}
   \item{stolen}{Frequency count, numeric}
   \item{wallet}{Frequency count, numeric}
   \item{til}{Frequency count, numeric}
   \item{repair}{Frequency count, numeric}
   \item{tuesday}{Frequency count, numeric}
   \item{fit}{Frequency count, numeric}
   \item{nonsens}{Frequency count, numeric}
   \item{incur}{Frequency count, numeric}
   \item{abt}{Frequency count, numeric}
   \item{deni}{Frequency count, numeric}
   \item{prove}{Frequency count, numeric}
   \item{yrs}{Frequency count, numeric}
   \item{etc}{Frequency count, numeric}
   \item{investig}{Frequency count, numeric}
   \item{sell}{Frequency count, numeric}
   \item{cuz}{Frequency count, numeric}
   \item{hubbi}{Frequency count, numeric}
   \item{power}{Frequency count, numeric}
   \item{rare}{Frequency count, numeric}
   \item{poorcustomerservic}{Frequency count, numeric}
   \item{forth}{Frequency count, numeric}
   \item{multipl}{Frequency count, numeric}
   \item{angri}{Frequency count, numeric}
   \item{lake}{Frequency count, numeric}
   \item{consist}{Frequency count, numeric}
   \item{commit}{Frequency count, numeric}
   \item{merger}{Frequency count, numeric}
   \item{glitch}{Frequency count, numeric}
   \item{mainten}{Frequency count, numeric}
   \item{state}{Frequency count, numeric}
   \item{truth}{Frequency count, numeric}
   \item{unfriend}{Frequency count, numeric}
   \item{ill}{Frequency count, numeric}
   \item{level}{Frequency count, numeric}
   \item{lufthansa}{Frequency count, numeric}
   \item{replac}{Frequency count, numeric}
   \item{sick}{Frequency count, numeric}
   \item{typic}{Frequency count, numeric}
   \item{prompt}{Frequency count, numeric}
   \item{simpli}{Frequency count, numeric}
   \item{twice}{Frequency count, numeric}
   \item{hate}{Frequency count, numeric}
   \item{orf}{Frequency count, numeric}
   \item{driven}{Frequency count, numeric}
   \item{uauau}{Frequency count, numeric}
   \item{age}{Frequency count, numeric}
   \item{choos}{Frequency count, numeric}
   \item{poorservic}{Frequency count, numeric}
   \item{general}{Frequency count, numeric}
   \item{attent}{Frequency count, numeric}
   \item{lit}{Frequency count, numeric}
   \item{aspen}{Frequency count, numeric}
   \item{certain}{Frequency count, numeric}
   \item{florida}{Frequency count, numeric}
   \item{recov}{Frequency count, numeric}
   \item{allow}{Frequency count, numeric}
   \item{overweight}{Frequency count, numeric}
   \item{measur}{Frequency count, numeric}
   \item{woman}{Frequency count, numeric}
   \item{miser}{Frequency count, numeric}
   \item{block}{Frequency count, numeric}
   \item{dming}{Frequency count, numeric}
   \item{articl}{Frequency count, numeric}
   \item{redeem}{Frequency count, numeric}
   \item{ton}{Frequency count, numeric}
   \item{what}{Frequency count, numeric}
   \item{gain}{Frequency count, numeric}
   \item{musician}{Frequency count, numeric}
   \item{holder}{Frequency count, numeric}
   \item{whose}{Frequency count, numeric}
   \item{blog}{Frequency count, numeric}
   \item{cleveland}{Frequency count, numeric}
   \item{sunday}{Frequency count, numeric}
   \item{basic}{Frequency count, numeric}
   \item{assign}{Frequency count, numeric}
   \item{sens}{Frequency count, numeric}
   \item{doesnut}{Frequency count, numeric}
   \item{english}{Frequency count, numeric}
   \item{hassl}{Frequency count, numeric}
   \item{badservic}{Frequency count, numeric}
   \item{whatev}{Frequency count, numeric}
   \item{london}{Frequency count, numeric}
   \item{cute}{Frequency count, numeric}
   \item{kind}{Frequency count, numeric}
   \item{speed}{Frequency count, numeric}
   \item{cowork}{Frequency count, numeric}
   \item{hell}{Frequency count, numeric}
   \item{flightst}{Frequency count, numeric}
   \item{settl}{Frequency count, numeric}
   \item{rdu}{Frequency count, numeric}
   \item{shop}{Frequency count, numeric}
   \item{onboard}{Frequency count, numeric}
   \item{pts}{Frequency count, numeric}
   \item{compar}{Frequency count, numeric}
   \item{mex}{Frequency count, numeric}
   \item{ten}{Frequency count, numeric}
   \item{usairway}{Frequency count, numeric}
   \item{usairwaysfail}{Frequency count, numeric}
   \item{practic}{Frequency count, numeric}
   \item{occur}{Frequency count, numeric}
   \item{outsourc}{Frequency count, numeric}
   \item{collect}{Frequency count, numeric}
   \item{suggest}{Frequency count, numeric}
   \item{formal}{Frequency count, numeric}
   \item{handler}{Frequency count, numeric}
   \item{bridg}{Frequency count, numeric}
   \item{freez}{Frequency count, numeric}
   \item{temp}{Frequency count, numeric}
   \item{foot}{Frequency count, numeric}
   \item{broke}{Frequency count, numeric}
   \item{global}{Frequency count, numeric}
   \item{yyz}{Frequency count, numeric}
   \item{piss}{Frequency count, numeric}
   \item{engin}{Frequency count, numeric}
   \item{deplan}{Frequency count, numeric}
   \item{releas}{Frequency count, numeric}
   \item{advertis}{Frequency count, numeric}
   \item{extend}{Frequency count, numeric}
   \item{aircanada}{Frequency count, numeric}
   \item{lesson}{Frequency count, numeric}
   \item{written}{Frequency count, numeric}
   \item{pray}{Frequency count, numeric}
   \item{chair}{Frequency count, numeric}
   \item{meal}{Frequency count, numeric}
   \item{unreal}{Frequency count, numeric}
   \item{costum}{Frequency count, numeric}
   \item{understaf}{Frequency count, numeric}
   \item{beer}{Frequency count, numeric}
   \item{understood}{Frequency count, numeric}
   \item{oop}{Frequency count, numeric}
   \item{crazi}{Frequency count, numeric}
   \item{rob}{Frequency count, numeric}
   \item{shut}{Frequency count, numeric}
   \item{worthless}{Frequency count, numeric}
   \item{polic}{Frequency count, numeric}
   \item{clean}{Frequency count, numeric}
   \item{dri}{Frequency count, numeric}
   \item{msg}{Frequency count, numeric}
   \item{wet}{Frequency count, numeric}
   \item{piec}{Frequency count, numeric}
   \item{print}{Frequency count, numeric}
   \item{airway}{Frequency count, numeric}
   \item{mci}{Frequency count, numeric}
   \item{segment}{Frequency count, numeric}
   \item{slight}{Frequency count, numeric}
   \item{acct}{Frequency count, numeric}
   \item{eta}{Frequency count, numeric}
   \item{btv}{Frequency count, numeric}
   \item{cun}{Frequency count, numeric}
   \item{nite}{Frequency count, numeric}
   \item{worstcustomerservic}{Frequency count, numeric}
   \item{rerout}{Frequency count, numeric}
   \item{cri}{Frequency count, numeric}
   \item{garbag}{Frequency count, numeric}
   \item{epic}{Frequency count, numeric}
   \item{jetway}{Frequency count, numeric}
   \item{reagan}{Frequency count, numeric}
   \item{ber}{Frequency count, numeric}
   \item{till}{Frequency count, numeric}
   \item{frankfurt}{Frequency count, numeric}
   \item{fan}{Frequency count, numeric}
   \item{doubl}{Frequency count, numeric}
   \item{lisa}{Frequency count, numeric}
   \item{breakfast}{Frequency count, numeric}
   \item{irrit}{Frequency count, numeric}
   \item{humor}{Frequency count, numeric}
   \item{copi}{Frequency count, numeric}
   \item{ual}{Frequency count, numeric}
   \item{tampa}{Frequency count, numeric}
   \item{robot}{Frequency count, numeric}
   \item{visit}{Frequency count, numeric}
   \item{wear}{Frequency count, numeric}
   \item{grr}{Frequency count, numeric}
   \item{raleigh}{Frequency count, numeric}
   \item{terri}{Frequency count, numeric}
   \item{freq}{Frequency count, numeric}
   \item{rsw}{Frequency count, numeric}
   \item{lousi}{Frequency count, numeric}
   \item{fault}{Frequency count, numeric}
   \item{honeymoon}{Frequency count, numeric}
   \item{gotten}{Frequency count, numeric}
   \item{epicfail}{Frequency count, numeric}
   \item{awar}{Frequency count, numeric}
   \item{hilari}{Frequency count, numeric}
   \item{ktn}{Frequency count, numeric}
   \item{wreck}{Frequency count, numeric}
   \item{offici}{Frequency count, numeric}
   \item{pleasant}{Frequency count, numeric}
   \item{acknowledg}{Frequency count, numeric}
   \item{dot}{Frequency count, numeric}
   \item{brother}{Frequency count, numeric}
   \item{hous}{Frequency count, numeric}
   \item{pleasur}{Frequency count, numeric}
   \item{bdl}{Frequency count, numeric}
   \item{explor}{Frequency count, numeric}
   \item{facebook}{Frequency count, numeric}
   \item{tue}{Frequency count, numeric}
   \item{buf}{Frequency count, numeric}
   \item{downgrad}{Frequency count, numeric}
   \item{nashvill}{Frequency count, numeric}
   \item{surpris}{Frequency count, numeric}
   \item{freak}{Frequency count, numeric}
   \item{snowi}{Frequency count, numeric}
   \item{ineffici}{Frequency count, numeric}
   \item{tini}{Frequency count, numeric}
   \item{csr}{Frequency count, numeric}
   \item{dozen}{Frequency count, numeric}
   \item{kiosk}{Frequency count, numeric}
   \item{patienc}{Frequency count, numeric}
   \item{bus}{Frequency count, numeric}
   \item{prob}{Frequency count, numeric}
   \item{disgrac}{Frequency count, numeric}
   \item{hit}{Frequency count, numeric}
   \item{forev}{Frequency count, numeric}
   \item{tone}{Frequency count, numeric}
   \item{worker}{Frequency count, numeric}
   \item{lazi}{Frequency count, numeric}
   \item{frozen}{Frequency count, numeric}
   \item{metal}{Frequency count, numeric}
   \item{truck}{Frequency count, numeric}
   \item{valu}{Frequency count, numeric}
   \item{fast}{Frequency count, numeric}
   \item{reconsid}{Frequency count, numeric}
   \item{attitud}{Frequency count, numeric}
   \item{respect}{Frequency count, numeric}
   \item{pat}{Frequency count, numeric}
   \item{disgust}{Frequency count, numeric}
   \item{restor}{Frequency count, numeric}
   \item{worstairlineev}{Frequency count, numeric}
   \item{valuabl}{Frequency count, numeric}
   \item{park}{Frequency count, numeric}
   \item{desper}{Frequency count, numeric}
   \item{girlfriend}{Frequency count, numeric}
   \item{fraud}{Frequency count, numeric}
   \item{pretzel}{Frequency count, numeric}
   \item{lower}{Frequency count, numeric}
   \item{crappi}{Frequency count, numeric}
   \item{legal}{Frequency count, numeric}
   \item{hang}{Frequency count, numeric}
   \item{intent}{Frequency count, numeric}
   \item{felt}{Frequency count, numeric}
   \item{higher}{Frequency count, numeric}
   \item{qualiti}{Frequency count, numeric}
   \item{spanish}{Frequency count, numeric}
   \item{mech}{Frequency count, numeric}
   \item{retriev}{Frequency count, numeric}
   \item{except}{Frequency count, numeric}
   \item{contract}{Frequency count, numeric}
   \item{pvd}{Frequency count, numeric}
   \item{upcom}{Frequency count, numeric}
   \item{cover}{Frequency count, numeric}
   \item{dumb}{Frequency count, numeric}
   \item{hurt}{Frequency count, numeric}
   \item{won}{Frequency count, numeric}
   \item{memori}{Frequency count, numeric}
   \item{circumst}{Frequency count, numeric}
   \item{disconnect}{Frequency count, numeric}
   \item{reliabl}{Frequency count, numeric}
   \item{none}{Frequency count, numeric}
   \item{relat}{Frequency count, numeric}
   \item{act}{Frequency count, numeric}
   \item{organ}{Frequency count, numeric}
   \item{bay}{Frequency count, numeric}
   \item{import}{Frequency count, numeric}
   \item{describ}{Frequency count, numeric}
   \item{flightawar}{Frequency count, numeric}
   \item{per}{Frequency count, numeric}
   \item{tail}{Frequency count, numeric}
   \item{dinner}{Frequency count, numeric}
   \item{honest}{Frequency count, numeric}
   \item{capt}{Frequency count, numeric}
   \item{explan}{Frequency count, numeric}
   \item{cab}{Frequency count, numeric}
   \item{pit}{Frequency count, numeric}
   \item{dont}{Frequency count, numeric}
   \item{cheap}{Frequency count, numeric}
   \item{worstairlin}{Frequency count, numeric}
   \item{bummer}{Frequency count, numeric}
   \item{content}{Frequency count, numeric}
   \item{realiz}{Frequency count, numeric}
   \item{discov}{Frequency count, numeric}
   \item{ive}{Frequency count, numeric}
   \item{bhm}{Frequency count, numeric}
   \item{god}{Frequency count, numeric}
   \item{educ}{Frequency count, numeric}
   \item{volunt}{Frequency count, numeric}
   \item{philippin}{Frequency count, numeric}
   \item{aisl}{Frequency count, numeric}
   \item{drunk}{Frequency count, numeric}
   \item{log}{Frequency count, numeric}
   \item{strike}{Frequency count, numeric}
   \item{social}{Frequency count, numeric}
   \item{planet}{Frequency count, numeric}
   \item{hdn}{Frequency count, numeric}
   \item{concours}{Frequency count, numeric}
   \item{shuttl}{Frequency count, numeric}
   \item{known}{Frequency count, numeric}
   \item{sight}{Frequency count, numeric}
   \item{trap}{Frequency count, numeric}
   \item{numer}{Frequency count, numeric}
   \item{slept}{Frequency count, numeric}
   \item{prior}{Frequency count, numeric}
   \item{runway}{Frequency count, numeric}
   \item{fals}{Frequency count, numeric}
   \item{idiot}{Frequency count, numeric}
   \item{relationship}{Frequency count, numeric}
   \item{everytim}{Frequency count, numeric}
   \item{green}{Frequency count, numeric}
   \item{present}{Frequency count, numeric}
   \item{faa}{Frequency count, numeric}
   \item{suit}{Frequency count, numeric}
   \item{elit}{Frequency count, numeric}
   \item{staralli}{Frequency count, numeric}
   \item{sister}{Frequency count, numeric}
   \item{true}{Frequency count, numeric}
   \item{live}{Frequency count, numeric}
   \item{five}{Frequency count, numeric}
   \item{push}{Frequency count, numeric}
   \item{ass}{Frequency count, numeric}
   \item{sometim}{Frequency count, numeric}
   \item{stt}{Frequency count, numeric}
   \item{separ}{Frequency count, numeric}
   \item{often}{Frequency count, numeric}
   \item{bathroom}{Frequency count, numeric}
   \item{unitedsuck}{Frequency count, numeric}
   \item{toddler}{Frequency count, numeric}
   \item{competitor}{Frequency count, numeric}
   \item{pend}{Frequency count, numeric}
   \item{toilet}{Frequency count, numeric}
   \item{outrag}{Frequency count, numeric}
   \item{chi}{Frequency count, numeric}
   \item{safe}{Frequency count, numeric}
   \item{proof}{Frequency count, numeric}
   \item{automat}{Frequency count, numeric}
   \item{cater}{Frequency count, numeric}
   \item{tvs}{Frequency count, numeric}
   \item{overs}{Frequency count, numeric}
   \item{rent}{Frequency count, numeric}
   \item{cluster}{Frequency count, numeric}
   \item{confer}{Frequency count, numeric}
   \item{memphi}{Frequency count, numeric}
   \item{notifi}{Frequency count, numeric}
   \item{bereav}{Frequency count, numeric}
   \item{discount}{Frequency count, numeric}
   \item{funer}{Frequency count, numeric}
   \item{meant}{Frequency count, numeric}
   \item{passport}{Frequency count, numeric}
   \item{approach}{Frequency count, numeric}
   \item{rapid}{Frequency count, numeric}
   \item{tick}{Frequency count, numeric}
   \item{appal}{Frequency count, numeric}
   \item{journey}{Frequency count, numeric}
   \item{spell}{Frequency count, numeric}
   \item{perhap}{Frequency count, numeric}
   \item{ahold}{Frequency count, numeric}
   \item{delet}{Frequency count, numeric}
   \item{notch}{Frequency count, numeric}
   \item{cvg}{Frequency count, numeric}
   \item{shoulder}{Frequency count, numeric}
   \item{skill}{Frequency count, numeric}
   \item{allianc}{Frequency count, numeric}
   \item{belong}{Frequency count, numeric}
   \item{depend}{Frequency count, numeric}
   \item{star}{Frequency count, numeric}
   \item{staf}{Frequency count, numeric}
   \item{hospit}{Frequency count, numeric}
   \item{juan}{Frequency count, numeric}
   \item{personnel}{Frequency count, numeric}
   \item{particular}{Frequency count, numeric}
   \item{fam}{Frequency count, numeric}
   \item{pricey}{Frequency count, numeric}
   \item{unitedfail}{Frequency count, numeric}
   \item{inexcus}{Frequency count, numeric}
   \item{million}{Frequency count, numeric}
   \item{immedi}{Frequency count, numeric}
   \item{weight}{Frequency count, numeric}
   \item{arrang}{Frequency count, numeric}
   \item{condit}{Frequency count, numeric}
   \item{guarante}{Frequency count, numeric}
   \item{albani}{Frequency count, numeric}
   \item{peanut}{Frequency count, numeric}
   \item{clarifi}{Frequency count, numeric}
   \item{unknown}{Frequency count, numeric}
   \item{wed}{Frequency count, numeric}
   \item{ramp}{Frequency count, numeric}
   \item{cranki}{Frequency count, numeric}
   \item{hello}{Frequency count, numeric}
   \item{detroit}{Frequency count, numeric}
   \item{bush}{Frequency count, numeric}
   \item{cle}{Frequency count, numeric}
   \item{alright}{Frequency count, numeric}
   \item{especi}{Frequency count, numeric}
   \item{shitti}{Frequency count, numeric}
   \item{amount}{Frequency count, numeric}
   \item{tremend}{Frequency count, numeric}
   \item{averag}{Frequency count, numeric}
   \item{abysm}{Frequency count, numeric}
   \item{unbeliev}{Frequency count, numeric}
   \item{decid}{Frequency count, numeric}
   \item{mark}{Frequency count, numeric}
   \item{children}{Frequency count, numeric}
   \item{draw}{Frequency count, numeric}
   \item{roundtrip}{Frequency count, numeric}
   \item{loyalti}{Frequency count, numeric}
   \item{perk}{Frequency count, numeric}
   \item{download}{Frequency count, numeric}
   \item{hnl}{Frequency count, numeric}
   \item{cxl}{Frequency count, numeric}
   \item{progress}{Frequency count, numeric}
   \item{fiasco}{Frequency count, numeric}
   \item{callback}{Frequency count, numeric}
   \item{corpor}{Frequency count, numeric}
   \item{relax}{Frequency count, numeric}
   \item{mother}{Frequency count, numeric}
   \item{natur}{Frequency count, numeric}
   \item{wasnt}{Frequency count, numeric}
   \item{proud}{Frequency count, numeric}
   \item{water}{Frequency count, numeric}
   \item{legroom}{Frequency count, numeric}
   \item{rest}{Frequency count, numeric}
   \item{cooki}{Frequency count, numeric}
   \item{chose}{Frequency count, numeric}
   \item{young}{Frequency count, numeric}
   \item{tmrw}{Frequency count, numeric}
   \item{weird}{Frequency count, numeric}
   \item{stream}{Frequency count, numeric}
   \item{decent}{Frequency count, numeric}
   \item{merg}{Frequency count, numeric}
   \item{pound}{Frequency count, numeric}
   \item{fire}{Frequency count, numeric}
   \item{blow}{Frequency count, numeric}
   \item{eventu}{Frequency count, numeric}
   \item{pain}{Frequency count, numeric}
   \item{cnn}{Frequency count, numeric}
   \item{approv}{Frequency count, numeric}
   \item{flts}{Frequency count, numeric}
   \item{ship}{Frequency count, numeric}
   \item{similar}{Frequency count, numeric}
   \item{inch}{Frequency count, numeric}
   \item{couldnt}{Frequency count, numeric}
   \item{south}{Frequency count, numeric}
   \item{six}{Frequency count, numeric}
   \item{directtv}{Frequency count, numeric}
   \item{advic}{Frequency count, numeric}
   \item{closer}{Frequency count, numeric}
   \item{traffic}{Frequency count, numeric}
   \item{beliz}{Frequency count, numeric}
   \item{unlik}{Frequency count, numeric}
   \item{congratul}{Frequency count, numeric}
   \item{airbus}{Frequency count, numeric}
   \item{wall}{Frequency count, numeric}
   \item{shout}{Frequency count, numeric}
   \item{disregard}{Frequency count, numeric}
   \item{dragon}{Frequency count, numeric}
   \item{pen}{Frequency count, numeric}
   \item{backup}{Frequency count, numeric}
   \item{remain}{Frequency count, numeric}
   \item{bna}{Frequency count, numeric}
   \item{jax}{Frequency count, numeric}
   \item{exit}{Frequency count, numeric}
   \item{wheelchair}{Frequency count, numeric}
   \item{plenti}{Frequency count, numeric}
   \item{yay}{Frequency count, numeric}
   \item{fra}{Frequency count, numeric}
   \item{die}{Frequency count, numeric}
   \item{plastic}{Frequency count, numeric}
   \item{stock}{Frequency count, numeric}
   \item{roc}{Frequency count, numeric}
   \item{orlean}{Frequency count, numeric}
   \item{map}{Frequency count, numeric}
   \item{insult}{Frequency count, numeric}
   \item{north}{Frequency count, numeric}
   \item{sale}{Frequency count, numeric}
   \item{tkt}{Frequency count, numeric}
   \item{smile}{Frequency count, numeric}
   \item{ind}{Frequency count, numeric}
   \item{interview}{Frequency count, numeric}
   \item{biggest}{Frequency count, numeric}
   \item{ripoff}{Frequency count, numeric}
   \item{although}{Frequency count, numeric}
   \item{difficulti}{Frequency count, numeric}
   \item{nothappi}{Frequency count, numeric}
   \item{celebr}{Frequency count, numeric}
   \item{slc}{Frequency count, numeric}
   \item{argu}{Frequency count, numeric}
   \item{histori}{Frequency count, numeric}
   \item{overal}{Frequency count, numeric}
   \item{expedia}{Frequency count, numeric}
   \item{document}{Frequency count, numeric}
   \item{technic}{Frequency count, numeric}
   \item{login}{Frequency count, numeric}
   \item{incent}{Frequency count, numeric}
   \item{bahama}{Frequency count, numeric}
   \item{horrend}{Frequency count, numeric}
   \item{procedur}{Frequency count, numeric}
   \item{parti}{Frequency count, numeric}
   \item{cell}{Frequency count, numeric}
   \item{teach}{Frequency count, numeric}
   \item{isnt}{Frequency count, numeric}
   \item{tpa}{Frequency count, numeric}
   \item{msp}{Frequency count, numeric}
   \item{chariti}{Frequency count, numeric}
   \item{aviat}{Frequency count, numeric}
   \item{cup}{Frequency count, numeric}
   \item{extens}{Frequency count, numeric}
   \item{somebodi}{Frequency count, numeric}
   \item{john}{Frequency count, numeric}
   \item{verifi}{Frequency count, numeric}
   \item{furious}{Frequency count, numeric}
   \item{bankrupt}{Frequency count, numeric}
   \item{iuv}{Frequency count, numeric}
   \item{type}{Frequency count, numeric}
   \item{googl}{Frequency count, numeric}
   \item{cart}{Frequency count, numeric}
   \item{golf}{Frequency count, numeric}
   \item{signific}{Frequency count, numeric}
   \item{insan}{Frequency count, numeric}
   \item{remot}{Frequency count, numeric}
   \item{volum}{Frequency count, numeric}
   \item{patient}{Frequency count, numeric}
   \item{folder}{Frequency count, numeric}
   \item{darn}{Frequency count, numeric}
   \item{tight}{Frequency count, numeric}
   \item{clarif}{Frequency count, numeric}
   \item{east}{Frequency count, numeric}
   \item{embarrass}{Frequency count, numeric}
   \item{warn}{Frequency count, numeric}
   \item{wednesday}{Frequency count, numeric}
   \item{combin}{Frequency count, numeric}
   \item{res}{Frequency count, numeric}
   \item{snack}{Frequency count, numeric}
   \item{belt}{Frequency count, numeric}
   \item{autom}{Frequency count, numeric}
   \item{transport}{Frequency count, numeric}
   \item{mad}{Frequency count, numeric}
   \item{june}{Frequency count, numeric}
   \item{specif}{Frequency count, numeric}
   \item{function}{Frequency count, numeric}
   \item{lavatori}{Frequency count, numeric}
   \item{winner}{Frequency count, numeric}
   \item{roll}{Frequency count, numeric}
   \item{accur}{Frequency count, numeric}
   \item{companion}{Frequency count, numeric}
   \item{necessari}{Frequency count, numeric}
   \item{thousand}{Frequency count, numeric}
   \item{pbi}{Frequency count, numeric}
   \item{father}{Frequency count, numeric}
   \item{lifetim}{Frequency count, numeric}
   \item{burn}{Frequency count, numeric}
   \item{misplac}{Frequency count, numeric}
   \item{laguardia}{Frequency count, numeric}
   \item{size}{Frequency count, numeric}
   \item{box}{Frequency count, numeric}
   \item{haha}{Frequency count, numeric}
   \item{donat}{Frequency count, numeric}
   \item{wth}{Frequency count, numeric}
   \item{hungri}{Frequency count, numeric}
   \item{maint}{Frequency count, numeric}
   \item{firstclass}{Frequency count, numeric}
   \item{massiv}{Frequency count, numeric}
   \item{fantast}{Frequency count, numeric}
   \item{annual}{Frequency count, numeric}
   \item{exchang}{Frequency count, numeric}
   \item{cont}{Frequency count, numeric}
   \item{dial}{Frequency count, numeric}
   \item{tablet}{Frequency count, numeric}
   \item{courteous}{Frequency count, numeric}
   \item{yup}{Frequency count, numeric}
   \item{alert}{Frequency count, numeric}
   \item{fave}{Frequency count, numeric}
   \item{dest}{Frequency count, numeric}
   \item{straight}{Frequency count, numeric}
   \item{anniversari}{Frequency count, numeric}
   \item{expedit}{Frequency count, numeric}
   \item{awe}{Frequency count, numeric}
   \item{buffalo}{Frequency count, numeric}
   \item{estim}{Frequency count, numeric}
   \item{join}{Frequency count, numeric}
   \item{gas}{Frequency count, numeric}
   \item{sept}{Frequency count, numeric}
   \item{magic}{Frequency count, numeric}
   \item{nasti}{Frequency count, numeric}
   \item{admir}{Frequency count, numeric}
   \item{customerservicefail}{Frequency count, numeric}
   \item{heavi}{Frequency count, numeric}
   \item{incom}{Frequency count, numeric}
   \item{stick}{Frequency count, numeric}
   \item{survey}{Frequency count, numeric}
   \item{rant}{Frequency count, numeric}
   \item{puerto}{Frequency count, numeric}
   \item{rico}{Frequency count, numeric}
   \item{annricord}{Frequency count, numeric}
   \item{alot}{Frequency count, numeric}
   \item{pet}{Frequency count, numeric}
   \item{west}{Frequency count, numeric}
   \item{admit}{Frequency count, numeric}
   \item{boss}{Frequency count, numeric}
   \item{women}{Frequency count, numeric}
   \item{waiver}{Frequency count, numeric}
   \item{incorrect}{Frequency count, numeric}
   \item{dobett}{Frequency count, numeric}
   \item{weigh}{Frequency count, numeric}
   \item{whether}{Frequency count, numeric}
   \item{wrote}{Frequency count, numeric}
   \item{custserv}{Frequency count, numeric}
   \item{parent}{Frequency count, numeric}
   \item{term}{Frequency count, numeric}
   \item{camera}{Frequency count, numeric}
   \item{didnut}{Frequency count, numeric}
   \item{pittsburgh}{Frequency count, numeric}
   \item{tab}{Frequency count, numeric}
   \item{harder}{Frequency count, numeric}
   \item{van}{Frequency count, numeric}
   \item{somewher}{Frequency count, numeric}
   \item{rush}{Frequency count, numeric}
   \item{letter}{Frequency count, numeric}
   \item{fresh}{Frequency count, numeric}
   \item{mission}{Frequency count, numeric}
   \item{subpar}{Frequency count, numeric}
   \item{ric}{Frequency count, numeric}
   \item{your}{Frequency count, numeric}
   \item{august}{Frequency count, numeric}
   \item{reward}{Frequency count, numeric}
   \item{headach}{Frequency count, numeric}
   \item{conflict}{Frequency count, numeric}
   \item{pref}{Frequency count, numeric}
   \item{fort}{Frequency count, numeric}
   \item{invit}{Frequency count, numeric}
   \item{flytpa}{Frequency count, numeric}
   \item{protect}{Frequency count, numeric}
   \item{neither}{Frequency count, numeric}
   \item{laugh}{Frequency count, numeric}
   \item{scale}{Frequency count, numeric}
   \item{nonexist}{Frequency count, numeric}
   \item{blah}{Frequency count, numeric}
   \item{tear}{Frequency count, numeric}
   \item{hartford}{Frequency count, numeric}
   \item{mke}{Frequency count, numeric}
   \item{warm}{Frequency count, numeric}
   \item{odd}{Frequency count, numeric}
   \item{packag}{Frequency count, numeric}
   \item{omaha}{Frequency count, numeric}
   \item{initi}{Frequency count, numeric}
   \item{lauderdal}{Frequency count, numeric}
   \item{lmao}{Frequency count, numeric}
   \item{feed}{Frequency count, numeric}
   \item{earth}{Frequency count, numeric}
   \item{fortun}{Frequency count, numeric}
   \item{heart}{Frequency count, numeric}
   \item{dang}{Frequency count, numeric}
   \item{ici}{Frequency count, numeric}
   \item{magazin}{Frequency count, numeric}
   \item{weur}{Frequency count, numeric}
   \item{greatservic}{Frequency count, numeric}
   \item{rno}{Frequency count, numeric}
   \item{street}{Frequency count, numeric}
   \item{nexus}{Frequency count, numeric}
   \item{cheat}{Frequency count, numeric}
   \item{inde}{Frequency count, numeric}
   \item{there}{Frequency count, numeric}
   \item{qualifi}{Frequency count, numeric}
   \item{asham}{Frequency count, numeric}
   \item{speedi}{Frequency count, numeric}
   \item{success}{Frequency count, numeric}
   \item{gogo}{Frequency count, numeric}
   \item{hub}{Frequency count, numeric}
   \item{dime}{Frequency count, numeric}
   \item{disorgan}{Frequency count, numeric}
   \item{sigh}{Frequency count, numeric}
   \item{proactiv}{Frequency count, numeric}
   \item{phx}{Frequency count, numeric}
   \item{jacksonvill}{Frequency count, numeric}
   \item{drove}{Frequency count, numeric}
   \item{birmingham}{Frequency count, numeric}
   \item{brutal}{Frequency count, numeric}
   \item{midway}{Frequency count, numeric}
   \item{swa}{Frequency count, numeric}
   \item{ivr}{Frequency count, numeric}
   \item{proceed}{Frequency count, numeric}
   \item{spiritairlin}{Frequency count, numeric}
   \item{data}{Frequency count, numeric}
   \item{southwestairlin}{Frequency count, numeric}
   \item{phlairport}{Frequency count, numeric}
   \item{charleston}{Frequency count, numeric}
   \item{throughout}{Frequency count, numeric}
   \item{birth}{Frequency count, numeric}
   \item{gear}{Frequency count, numeric}
   \item{nonprofit}{Frequency count, numeric}
   \item{partnership}{Frequency count, numeric}
   \item{mdw}{Frequency count, numeric}
   \item{fell}{Frequency count, numeric}
   \item{destinationdragon}{Frequency count, numeric}
   \item{uisouthwestair}{Frequency count, numeric}
   \item{imaginedragon}{Frequency count, numeric}
   \item{voic}{Frequency count, numeric}
   \item{profession}{Frequency count, numeric}
   \item{unreli}{Frequency count, numeric}
   \item{test}{Frequency count, numeric}
   \item{featur}{Frequency count, numeric}
   \item{bae}{Frequency count, numeric}
   \item{contest}{Frequency count, numeric}
   \item{spirit}{Frequency count, numeric}
   \item{hunt}{Frequency count, numeric}
   \item{scaveng}{Frequency count, numeric}
   \item{injuri}{Frequency count, numeric}
   \item{hero}{Frequency count, numeric}
   \item{grate}{Frequency count, numeric}
   \item{est}{Frequency count, numeric}
   \item{reflect}{Frequency count, numeric}
   \item{station}{Frequency count, numeric}
   \item{antonio}{Frequency count, numeric}
   \item{noon}{Frequency count, numeric}
   \item{ohio}{Frequency count, numeric}
   \item{aug}{Frequency count, numeric}
   \item{columbus}{Frequency count, numeric}
   \item{oakland}{Frequency count, numeric}
   \item{bostonlogan}{Frequency count, numeric}
   \item{heavili}{Frequency count, numeric}
   \item{enrout}{Frequency count, numeric}
   \item{soak}{Frequency count, numeric}
   \item{rethink}{Frequency count, numeric}
   \item{stroller}{Frequency count, numeric}
   \item{athlet}{Frequency count, numeric}
   \item{aka}{Frequency count, numeric}
   \item{telephon}{Frequency count, numeric}
   \item{band}{Frequency count, numeric}
   \item{ring}{Frequency count, numeric}
   \item{color}{Frequency count, numeric}
   \item{compass}{Frequency count, numeric}
   \item{lucki}{Frequency count, numeric}
   \item{jamaica}{Frequency count, numeric}
   \item{loss}{Frequency count, numeric}
   \item{cali}{Frequency count, numeric}
   \item{sunni}{Frequency count, numeric}
   \item{harbor}{Frequency count, numeric}
   \item{batteri}{Frequency count, numeric}
   \item{baltimor}{Frequency count, numeric}
   \item{greenvill}{Frequency count, numeric}
   \item{usair}{Frequency count, numeric}
   \item{indi}{Frequency count, numeric}
   \item{northeast}{Frequency count, numeric}
   \item{cmon}{Frequency count, numeric}
   \item{loos}{Frequency count, numeric}
   \item{spoken}{Frequency count, numeric}
   \item{heck}{Frequency count, numeric}
   \item{colleagu}{Frequency count, numeric}
   \item{resch}{Frequency count, numeric}
   \item{louisvill}{Frequency count, numeric}
   \item{blue}{Frequency count, numeric}
   \item{fortunemagazin}{Frequency count, numeric}
   \item{lunch}{Frequency count, numeric}
   \item{anxious}{Frequency count, numeric}
   \item{inclement}{Frequency count, numeric}
   \item{gari}{Frequency count, numeric}
   \item{kelli}{Frequency count, numeric}
   \item{stupid}{Frequency count, numeric}
   \item{summer}{Frequency count, numeric}
   \item{cash}{Frequency count, numeric}
   \item{crowd}{Frequency count, numeric}
   \item{utah}{Frequency count, numeric}
   \item{velourl}{Frequency count, numeric}
   \item{triflight}{Frequency count, numeric}
   \item{steal}{Frequency count, numeric}
   \item{selfi}{Frequency count, numeric}
   \item{paypal}{Frequency count, numeric}
   \item{nearbi}{Frequency count, numeric}
   \item{pwm}{Frequency count, numeric}
   \item{aww}{Frequency count, numeric}
   \item{colleg}{Frequency count, numeric}
   \item{physic}{Frequency count, numeric}
   \item{savannah}{Frequency count, numeric}
   \item{anticip}{Frequency count, numeric}
   \item{uuuu}{Frequency count, numeric}
   \item{lauren}{Frequency count, numeric}
   \item{lovedragonss}{Frequency count, numeric}
   \item{costa}{Frequency count, numeric}
   \item{rica}{Frequency count, numeric}
   \item{chao}{Frequency count, numeric}
   \item{unload}{Frequency count, numeric}
   \item{pump}{Frequency count, numeric}
   \item{evalu}{Frequency count, numeric}
   \item{predict}{Frequency count, numeric}
   \item{disney}{Frequency count, numeric}
   \item{dealt}{Frequency count, numeric}
   \item{fastest}{Frequency count, numeric}
   \item{involv}{Frequency count, numeric}
   \item{regular}{Frequency count, numeric}
   \item{lodg}{Frequency count, numeric}
   \item{caller}{Frequency count, numeric}
   \item{nation}{Frequency count, numeric}
   \item{sju}{Frequency count, numeric}
   \item{fleek}{Frequency count, numeric}
   \item{trueblu}{Frequency count, numeric}
   \item{uijetblu}{Frequency count, numeric}
   \item{plat}{Frequency count, numeric}
   \item{beach}{Frequency count, numeric}
   \item{flyfi}{Frequency count, numeric}
   \item{viequ}{Frequency count, numeric}
   \item{njb}{Frequency count, numeric}
   \item{jblu}{Frequency count, numeric}
   \item{stole}{Frequency count, numeric}
   \item{wsj}{Frequency count, numeric}
   \item{barbado}{Frequency count, numeric}
   \item{cuba}{Frequency count, numeric}
   \item{reli}{Frequency count, numeric}
   \item{agenc}{Frequency count, numeric}
   \item{mem}{Frequency count, numeric}
   \item{appeas}{Frequency count, numeric}
   \item{republican}{Frequency count, numeric}
   \item{waterburi}{Frequency count, numeric}
   \item{flyingitforward}{Frequency count, numeric}
   \item{mint}{Frequency count, numeric}
   \item{nantucket}{Frequency count, numeric}
   \item{glass}{Frequency count, numeric}
   \item{mosaic}{Frequency count, numeric}
   \item{goodwil}{Frequency count, numeric}
   \item{getittogeth}{Frequency count, numeric}
   \item{endless}{Frequency count, numeric}
   \item{brandmanc}{Frequency count, numeric}
   \item{hpn}{Frequency count, numeric}
   \item{prais}{Frequency count, numeric}
   \item{circl}{Frequency count, numeric}
   \item{journal}{Frequency count, numeric}
   \item{sxm}{Frequency count, numeric}
   \item{asleep}{Frequency count, numeric}
   \item{blueman}{Frequency count, numeric}
   \item{digit}{Frequency count, numeric}
   \item{flyitforward}{Frequency count, numeric}
   \item{crewmemb}{Frequency count, numeric}
   \item{liveri}{Frequency count, numeric}
   \item{bottl}{Frequency count, numeric}
   \item{divers}{Frequency count, numeric}
   \item{dissatisfi}{Frequency count, numeric}
   \item{comm}{Frequency count, numeric}
   \item{dump}{Frequency count, numeric}
   \item{paint}{Frequency count, numeric}
   \item{shoutout}{Frequency count, numeric}
   \item{alaska}{Frequency count, numeric}
   \item{weull}{Frequency count, numeric}
   \item{jvmchat}{Frequency count, numeric}
   \item{sympathi}{Frequency count, numeric}
   \item{dividend}{Frequency count, numeric}
   \item{chairman}{Frequency count, numeric}
   \item{uiusairway}{Frequency count, numeric}
   \item{exec}{Frequency count, numeric}
   \item{ctl}{Frequency count, numeric}
   \item{flydelta}{Frequency count, numeric}
   \item{logist}{Frequency count, numeric}
   \item{sexual}{Frequency count, numeric}
   \item{fear}{Frequency count, numeric}
   \item{aadvantag}{Frequency count, numeric}
   \item{cot}{Frequency count, numeric}
   \item{key}{Frequency count, numeric}
   \item{usairwayssuck}{Frequency count, numeric}
   \item{med}{Frequency count, numeric}
   \item{vile}{Frequency count, numeric}
   \item{uii}{Frequency count, numeric}
   \item{uiamericanair}{Frequency count, numeric}
   \item{assult}{Frequency count, numeric}
   \item{dfwairport}{Frequency count, numeric}
   \item{filmcrew}{Frequency count, numeric}
   \item{cowboycerron}{Frequency count, numeric}
   \item{americanview}{Frequency count, numeric}
}}
\source{
\url{https://www.figure-eight.com/data-for-everyone/}
}
\usage{
tweetsDTM
}
\description{
As the original source says,
}
\details{
"A sentiment analysis job about the problems of each major U.S. airline.
Twitter data was scraped from February of 2015 and contributors were
asked to first classify positive, negative, and neutral tweets, followed
by categorizing negative reasons (such as "late flight" or "rude service")."

The data provided here is an altered and cleaned version version of the original source.
}
\keyword{datasets}
