% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnb.update.R
\name{fnb.update}
\alias{fnb.update}
\alias{fnb.update.default}
\alias{fnb.update.fnb.bernoulli}
\title{Update function}
\usage{
fnb.update(object, x, y, sparse = FALSE, check = TRUE)

\method{fnb.update}{default}(object, x, y, sparse = FALSE,
  check = TRUE)

\method{fnb.update}{fnb.bernoulli}(object, x, y, sparse = FALSE,
  check = TRUE)
}
\arguments{
\item{object}{fitted model}

\item{x}{numeric matrix}

\item{y}{factor of classes}

\item{sparse}{boolean, set to true to use sparse matrices}

\item{check}{boolean, set to true to enable formal checks on input}
}
\description{
Update function
}
